refactor(commands): extract models, calendar, brief, enrich domains (gated, verbatim)
Pure MOVE, zero behavior change, the final production round of the God-file
split. Four domains lift out of commands/mod.rs into commands/{models,calendar,
brief,enrich}.rs (each #[path]-aliased to dodge E0255 vs crate::{transcribe::model,
calendar, brief_runner, enrich}); mod.rs 9,926 -> 7,531 (-2,395). lib.rs
generate_handler! BYTE-FOR-BYTE UNCHANGED.

30 #[tauri::command] moved (9 models + 3 calendar + 8 brief + 10 enrich); command
count conserved 74. Byte-identical: a scripted scan confirmed the ONLY lines in the
4 new files not verbatim-in-trunk are 10 fn->pub(crate) signature prepends, each
forced by a retained lifecycle_tests caller, bodies unchanged. No bare pub.

Gate/seal logic UNCHANGED, only relocated:
- brief/enrich gated readers keep meeting_is_unlocked / folder_is_unlocked ->
  AppError::Locked and list_meetings_visible / get_note_if_visible verbatim.
- every moved note WRITE keeps its upsert_note_reseal_if_locked seal-on-write seam;
  supersession apply keeps its per-row re-gate under the lifecycle guard verbatim.
- models/calendar are not content-gated (ML model/gateway mgmt + calendar connector).

Deliberately LEFT in mod.rs (shared, reached via super::): the note-write/seal web
(upsert_note_reseal_if_locked, overwrite_exported_note_guarded, meeting_is_unlocked,
lifecycle_guard, set_timeline_data_reseal_if_locked, index_wikilinks_best_effort,
auto_link_semantic_best_effort), the reindex/document-chunk helpers (startup repair
tick + documents.rs call them), the supersession gate helpers (analytics.rs calls
them), import_extraction_reasoner/user_memory_enabled (ask.rs/facts.rs).

Tests all stayed in mod.rs (8 #[path] test modules + 8 #[cfg(test)]).
plan_organize_notes/apply_organize_plan confirmed byte-identical. Verified: full
cargo test --lib 1992/0 (supersession 12, enrich 17, brief 21, model 103, gateway
51, reindex 14, note_assist 7, lock_read_gate 10, seal 145, reseal 1), clippy --lib
-D warnings clean, MSRV grep clean.
