+++ src-tauri/src/commands/mod.rs
@@ -211,6 +211,72 @@ pub use folders_commands::*;
 mod links_commands;
 pub use links_commands::*;
 
+// CALENDAR connector command surface (local macOS Calendar — NOT sealed content) — extracted
+// verbatim (God-file split, PURE MOVE — every body byte-identical, only relocated). These reads hit
+// the user's LOCAL calendar (osascript / the `meetnotes-calendar` EventKit sidecar), never sealed
+// meeting/note content, so there is NO content gate here (nothing to gate) and NO network egress
+// (any downstream cloud use of a `CalendarContext` still rides the `make_provider` firewall +
+// consent). Bound as `calendar_commands` (via `#[path]`) to avoid colliding with the crate-level
+// `crate::calendar` module these commands call (E0255). The glob re-export makes every moved command
+// resolve UNCHANGED at `crate::commands::…` for `generate_handler!` in `lib.rs` and every caller.
+#[path = "calendar.rs"]
+mod calendar_commands;
+pub use calendar_commands::*;
+
+// ML-MODEL + AI-GATEWAY management command surface (NOT content-gated) — extracted verbatim
+// (God-file split, PURE MOVE — every body byte-identical, only relocated). Two clusters, both
+// OUTSIDE the lock model: the AI-gateway/provider catalog + health probes (INBOUND-ONLY — at most an
+// `Authorization: Bearer` header leaves, never meeting content) and the model DOWNLOADS + vector
+// REINDEX trigger. `reindex_embeddings`'s visibility-gated corpus runs entirely through the SHARED
+// `reindex_embeddings_inner` (+ `backfill_document_chunks` / `index_document_row_kind_routed` and the
+// `ReindexResult` DTO), which all STAY in `commands/mod.rs` (also called by the startup repair tick +
+// the lifecycle tests); the moved command reaches them through `use super::*`. Bound as
+// `models_commands` (via `#[path]`) to avoid any name shadow with the crate-level model modules
+// (`crate::transcribe::model` / `crate::reason` / `crate::embed`). The glob re-export makes every
+// moved command resolve UNCHANGED at `crate::commands::…` for `generate_handler!` and every caller.
+#[path = "models.rs"]
+mod models_commands;
+pub use models_commands::*;
+
+// WEEKLY DIGEST + scheduled-BRIEFS command surface (a GATED domain where it reads content) —
+// extracted verbatim (God-file split, PURE MOVE — every body byte-identical, only relocated).
+// `generate_digest` builds its cloud corpus from VISIBLE meetings + VISIBLE notes only
+// (`list_meetings_visible` + `get_note_if_visible` — the same `visibility_clause` predicate MCP
+// uses), so a sealed-not-session-unlocked meeting's title/markdown never egress. `list_brief_runs`
+// returns pending proposals whose `note_md` was synthesized from VISIBLE-ONLY content AND is purged
+// on seal (`Db::purge_pending_brief_runs_tx`, inside the seal tx) — the documented no-per-meeting-gate
+// posture; the gate LOGIC is byte-identical, only relocated. `accept_brief` exports via
+// `crate::export::write_note` under the SHARED `vault_path` (which STAYS in `commands/mod.rs`,
+// reached via `super::`). Bound as `brief_commands` (via `#[path]`) to avoid any name shadow with the
+// crate-level `crate::brief_runner`. The glob re-export makes every moved command resolve UNCHANGED at
+// `crate::commands::…` for `generate_handler!` and every caller.
+#[path = "brief.rs"]
+mod brief_commands;
+pub use brief_commands::*;
+
+// NOTE-ENRICHMENT / VERIFY / SELECTION-ASSISTANT / SUPERSESSION / MEMORY-IMPORT command surface (a
+// GATED domain — every reader of meeting/note content keeps its gate, every note WRITE keeps its
+// write-gate + seal-on-write) — extracted verbatim (God-file split, PURE MOVE — every gate/seal/guard
+// body byte-identical, only relocated + visibility WIDENED to `pub(crate)` where a RETAINED lifecycle
+// test reaches a note-assist helper via `super::…`, the ONLY change to any moved item). The moved
+// readers keep `meeting_is_unlocked`/`folder_is_unlocked` (→ `AppError::Locked`) verbatim; the moved
+// note WRITES keep their seal-on-write seam (`upsert_note_reseal_if_locked` — a session-unlocked
+// LOCKED folder re-seals the fresh markdown; a SEALED note is refused). The supersession commands
+// RE-GATE each row at apply time (the prune↔seal TOCTOU discipline) under the shared lifecycle guard.
+// The SHARED note-write/seal/gate web (`upsert_note_reseal_if_locked`/`overwrite_exported_note_guarded`/
+// `meeting_is_unlocked`/`folder_is_unlocked`/`lifecycle_guard`/`set_timeline_data_reseal_if_locked`/
+// `index_wikilinks_best_effort`/`auto_link_semantic_best_effort`), the SUPERSESSION gate helpers
+// (`source_is_stampable`/`folder_locked_on_disk`/`note_file_for` — also called by the sibling
+// `commands/analytics.rs`), and the memory helpers (`import_extraction_reasoner`/`user_memory_enabled`,
+// owned by the siblings `ask.rs`/`facts.rs`) all STAY in `commands/mod.rs`; the moved commands reach
+// them through `use super::*`. Bound as `enrich_commands` (via `#[path]`) to avoid colliding with the
+// crate-level `crate::enrich` module these commands call (E0255). The glob re-export makes every moved
+// command resolve UNCHANGED at `crate::commands::…` for `generate_handler!` and every caller (incl.
+// the STAYING test modules).
+#[path = "enrich.rs"]
+mod enrich_commands;
+pub use enrich_commands::*;
+
 /// Keychain account for the AI Gateway API key (matches `summarize::GATEWAY_KEY_ACCOUNT`).
 /// Strictly separate from `ANTHROPIC_KEY_ACCOUNT` (defined in `commands/secrets.rs`) — never a
 /// fallback to the Anthropic key (R3). Kept here because the gateway model-listing helpers below
@@ -1467,337 +1533,6 @@ pub(crate) fn update_note_inner_with(
     })
 }
 
-/// VERIFY PASS (read-only): extract Jira issue keys from the meeting's note and check each against
-/// LIVE Jira. GATED: sealed-not-unlocked meetings refuse (a verify against a blanked note would be
-/// nonsense AND a read-gate bypass). Consent-gated: rides the Jira connector's enable+consent+key
-/// gate (fail-closed `NeedsConsent` maps to `AppError::Unavailable`). NEVER called proactively —
-/// FE-invoked only. Findings are computed against the note WITH OLD MARKERS STRIPPED so line
-/// numbers line up with `apply_verify_markers`' post-strip numbering.
-#[tauri::command]
-pub async fn verify_note_sources(
-    state: State<'_, AppState>,
-    meeting_id: String,
-) -> Result<Vec<crate::verify::VerifyFinding>, AppError> {
-    verify_note_sources_inner(state.inner(), meeting_id).await
-}
-
-pub(crate) async fn verify_note_sources_inner(
-    state: &AppState,
-    meeting_id: String,
-) -> Result<Vec<crate::verify::VerifyFinding>, AppError> {
-    if !meeting_is_unlocked(state, &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to verify the note".into(),
-        ));
-    }
-    // Brain v2 L5 — SESSION verify cache, checked AFTER the read gate (the gate is never skipped
-    // for a cache hit). A hit re-renders the panel without a second Jira egress; the cache is
-    // RAM-only and cleared on relock_folder / relock_all, so it never outlives the session unlock.
-    if let Some(cached) = state
-        .verify_cache
-        .lock()
-        .map_err(|_| AppError::Other(anyhow::anyhow!("verify cache lock")))?
-        .get(&meeting_id)
-    {
-        return Ok(cached.clone());
-    }
-    let note = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    // Strip our own old CALLOUT first (its body lines carry issue keys of their own), THEN the
-    // inline markers, so extraction/judgment sees the canonical note lines and line numbers line
-    // up with `apply_verify_markers`' post-strip numbering.
-    let base = crate::verify::apply_verify_callout(&note.markdown, &[], "");
-    let stripped = crate::verify::apply_verify_markers(&base, &[]);
-    let keys = crate::verify::extract_issue_keys(&stripped);
-    if keys.is_empty() {
-        return Ok(Vec::new());
-    }
-    let config = state
-        .config
-        .lock()
-        .map_err(|_| AppError::Other(anyhow::anyhow!("config lock")))?
-        .clone();
-    let registry = crate::connectors::ConnectorRegistry::build(&config);
-    let lines: Vec<&str> = stripped.lines().collect();
-    let mut findings = Vec::with_capacity(keys.len());
-    for (line_no, key) in keys {
-        let snap = registry.jira_lookup(&key).await.map_err(AppError::from)?;
-        let line_text = lines.get(line_no - 1).copied().unwrap_or("");
-        let (verdict, detail) = crate::verify::judge(line_text, &key, snap.as_ref());
-        let url = snap.map(|s| s.url).unwrap_or_default();
-        findings.push(crate::verify::VerifyFinding {
-            line_no,
-            key,
-            verdict,
-            detail,
-            url,
-        });
-    }
-    // Populate the session cache (RAM-only; cleared on relock). A poisoned lock only skips the
-    // cache — the findings still return.
-    if let Ok(mut cache) = state.verify_cache.lock() {
-        cache.insert(meeting_id, findings.clone());
-    }
-    Ok(findings)
-}
-
-/// Apply verify markers to the note (WRITE — same gate + save/re-export tail as `update_note`).
-/// Takes the findings the user just reviewed in the panel; validates every key's strict shape.
-#[tauri::command]
-pub fn apply_note_verify_markers(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    findings: Vec<crate::verify::VerifyFinding>,
-) -> Result<NoteDto, AppError> {
-    apply_note_verify_markers_inner(state.inner(), meeting_id, findings)
-}
-
-pub(crate) fn apply_note_verify_markers_inner(
-    state: &AppState,
-    meeting_id: String,
-    findings: Vec<crate::verify::VerifyFinding>,
-) -> Result<NoteDto, AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across gate+write so a
-    // concurrent relock/seal cannot land between the unlock check and the upsert.
-    let _lifecycle = lifecycle_guard(state);
-    if !meeting_is_unlocked(state, &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to edit the note".into(),
-        ));
-    }
-    for f in &findings {
-        let ok = crate::verify::extract_issue_keys(&f.key)
-            .first()
-            .map(|(_, k)| k == &f.key)
-            .unwrap_or(false);
-        if !ok {
-            return Err(AppError::InvalidArg("invalid issue key in findings".into()));
-        }
-    }
-    let existing = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    // Brain v2 L5 — strip our own old CALLOUT first (so `apply_verify_markers`' internal
-    // marker-strip numbering matches the findings, which were computed against the
-    // callout-stripped note), apply the inline markers, then append the fresh consolidated
-    // `> [!verify]-` callout dated now. All three regions are self-managed + idempotent.
-    let base = crate::verify::apply_verify_callout(&existing.markdown, &[], "");
-    let marked = crate::verify::apply_verify_markers(&base, &findings);
-    let as_of = chrono::Utc::now().to_rfc3339();
-    let marked = crate::verify::apply_verify_callout(&marked, &findings, &as_of);
-    // Save + re-export — the exact `update_note` tail, with `marked`. Persisting via the
-    // seal-on-write upsert keeps the callout in the CANONICAL DB note markdown so it SEALS with the
-    // note under the folder lock (the enrich.rs persistence lesson — and, for a session-unlocked
-    // LOCKED folder, the fresh markdown is re-sealed into `content_blob` in the same write); the
-    // vault `.md` re-export follows when one exists.
-    let created_at = chrono::Utc::now().to_rfc3339();
-    upsert_note_reseal_if_locked(
-        state,
-        &NoteRecord {
-            meeting_id: meeting_id.clone(),
-            provider_id: existing.provider_id.clone(),
-            markdown: marked.clone(),
-            created_at,
-            exported_path: existing.exported_path.clone(),
-            model_requested: existing.model_requested.clone(),
-            model_served: existing.model_served.clone(),
-            gateway_host: existing.gateway_host.clone(),
-        },
-    )?;
-    if let Some(path) = existing.exported_path.as_deref() {
-        overwrite_exported_note_guarded(state, &meeting_id, &existing.provider_id, path, &marked)?;
-    }
-    Ok(NoteDto {
-        meeting_id,
-        provider_id: existing.provider_id,
-        markdown: marked,
-        exported_path: existing.exported_path,
-    })
-}
-
-/// `enrich_note_context(meeting_id) -> Vec<ContextHit>` — CONNECTOR-AGNOSTIC preview of live context
-/// to fold into the note. Read side: gathers hits from EVERY exposed (enabled + consented + keyed)
-/// connector via the same registry the brain uses. Two modes (see the research brief):
-/// - **Identifier lookup** (precise, minimal egress): Jira issue keys already in the note → live
-///   `jira_lookup`. Only a validated `PROJ-123` leaves the Mac — never note content.
-/// - **Free-text search** (fuzzy): every OTHER exposed connector (Slack/web) is searched for the
-///   meeting's TITLE, through the framework's redaction + content-free egress ledger.
-///
-/// This is the EGRESS moment (an explicit user action, like `verify_note_sources`); the returned
-/// hits are reviewed in the FE and only WRITTEN by `apply_note_enrichment`. Lock-gated: a
-/// sealed-not-unlocked meeting refuses BEFORE any connector call. Empty vec = nothing to add.
-#[tauri::command]
-pub async fn enrich_note_context(
-    state: State<'_, AppState>,
-    meeting_id: String,
-) -> Result<Vec<crate::enrich::ContextHit>, AppError> {
-    enrich_note_context_inner(state.inner(), meeting_id).await
-}
-
-/// How many free-text search hits to keep per connector (bounds egress-result noise; the caller
-/// still reviews + can drop each before applying).
-const ENRICH_SEARCH_HITS_PER_CONNECTOR: usize = 3;
-
-pub(crate) async fn enrich_note_context_inner(
-    state: &AppState,
-    meeting_id: String,
-) -> Result<Vec<crate::enrich::ContextHit>, AppError> {
-    // READ-GATE FIRST — a sealed-not-unlocked meeting refuses before ANY connector egress.
-    if !meeting_is_unlocked(state, &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to add live context".into(),
-        ));
-    }
-    let note = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    // Clean base = the note with our own prior context block stripped, so key-extraction sees the
-    // canonical prose (never our appended callout).
-    let base = crate::enrich::apply_context_markers(&note.markdown, &[], "");
-    let title = state
-        .db
-        .get_meeting(&meeting_id)?
-        .and_then(|m| m.title)
-        .unwrap_or_default();
-
-    let config = state
-        .config
-        .lock()
-        .map_err(|_| AppError::Other(anyhow::anyhow!("config lock")))?
-        .clone();
-    let registry = crate::connectors::ConnectorRegistry::build(&config);
-
-    let mut hits: Vec<crate::enrich::ContextHit> = Vec::new();
-
-    // ── Identifier-lookup mode (precise): Jira issue keys → live status. Egresses only the key. ──
-    if registry.has("jira") {
-        for (_line, key) in crate::verify::extract_issue_keys(&base) {
-            if let Ok(Some(snap)) = registry.jira_lookup(&key).await {
-                let mut detail = format!("{} · {}", snap.key, snap.status);
-                if let Some(due) = snap.due.as_deref().filter(|d| !d.is_empty()) {
-                    detail.push_str(&format!(" · due {due}"));
-                }
-                if !snap.summary.is_empty() {
-                    detail.push_str(&format!(" — {}", snap.summary));
-                }
-                hits.push(crate::enrich::ContextHit {
-                    source: "Jira".to_string(),
-                    detail,
-                    url: Some(snap.url).filter(|u| !u.is_empty()),
-                });
-            }
-        }
-    }
-
-    // ── Free-text search mode (fuzzy): every OTHER exposed connector, queried on the meeting title.
-    // The query is redacted + ledgered by the registry; skip when there is no title to search on. ──
-    if !title.trim().is_empty() {
-        for id in registry.ids() {
-            if id == "jira" {
-                continue; // handled precisely above — never double-pull Jira.
-            }
-            if let Ok(results) = registry.search(id, &title).await {
-                for hit in results.into_iter().take(ENRICH_SEARCH_HITS_PER_CONNECTOR) {
-                    let detail = if hit.snippet.trim().is_empty() {
-                        hit.title
-                    } else {
-                        format!("{} — {}", hit.title, hit.snippet)
-                    };
-                    hits.push(crate::enrich::ContextHit {
-                        // Loud attribution from the connector itself (e.g. "Slack", "web · Brave").
-                        source: hit.source_label,
-                        detail,
-                        url: Some(hit.url).filter(|u| !u.is_empty()),
-                    });
-                }
-            }
-        }
-    }
-
-    Ok(hits)
-}
-
-/// `apply_note_enrichment(meeting_id, hits) -> NoteDto` — WRITE the reviewed context hits into the
-/// note as one consolidated `> [!context]-` callout (dated now), via the EXACT `update_note` save +
-/// re-export tail — so it persists in the CANONICAL DB note markdown and SEALS with the note under
-/// the folder lock (NOT the vault-file-only path). No egress here (the hits were fetched by
-/// `enrich_note_context`). Lock-gated. Passing an empty `hits` STRIPS the block (byte-exact undo).
-#[tauri::command]
-pub fn apply_note_enrichment(
-    state: State<'_, AppState>,
-    meeting_id: String,
-    hits: Vec<crate::enrich::ContextHit>,
-) -> Result<NoteDto, AppError> {
-    apply_note_enrichment_inner(state.inner(), meeting_id, hits)
-}
-
-pub(crate) fn apply_note_enrichment_inner(
-    state: &AppState,
-    meeting_id: String,
-    hits: Vec<crate::enrich::ContextHit>,
-) -> Result<NoteDto, AppError> {
-    // BLK-1 / TOCTOU: hold the lifecycle guard across the whole check-then-write so a concurrent seal
-    // cannot slip between the gate and the `upsert_note`+`overwrite_note` (same leak class Lane A's
-    // `link_related_notes_inner` guards). Lock order `lifecycle ⊃ db`.
-    let _lifecycle = lifecycle_guard(state);
-    if !meeting_is_unlocked(state, &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to edit the note".into(),
-        ));
-    }
-    // SEAL-SAFETY GATE (mirrors Lane A): a SEALED note (any provider row carries a content_blob) has a
-    // TRANSIENT `markdown` column — blanked on relock, restored from `content_blob` on unlock. Writing
-    // enriched markdown into it would be silently dropped on the next relock (content_blob is
-    // canonical), and — for the auto-file-into-locked case where the column is blank but the folder is
-    // session-unlocked — could re-materialize plaintext into a sealed note. So refuse enrichment on a
-    // sealed note even when the session has it unlocked.
-    let sealed = state
-        .db
-        .sealable_notes_for_meeting(&meeting_id)?
-        .iter()
-        .any(|n| n.content_blob.is_some());
-    if sealed {
-        return Err(AppError::Locked(
-            "this meeting's note is sealed — enrichment can't be persisted while locked".into(),
-        ));
-    }
-    let existing = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    let as_of = chrono::Utc::now().to_rfc3339();
-    let enriched = crate::enrich::apply_context_markers(&existing.markdown, &hits, &as_of);
-    let created_at = chrono::Utc::now().to_rfc3339();
-    // Seal-on-write seam (audit F1): the sealed case was refused above, but a LOCKED folder whose
-    // note was never sealed (auto-filed while session-unlocked) still re-seals the fresh markdown.
-    upsert_note_reseal_if_locked(
-        state,
-        &NoteRecord {
-            meeting_id: meeting_id.clone(),
-            provider_id: existing.provider_id.clone(),
-            markdown: enriched.clone(),
-            created_at,
-            exported_path: existing.exported_path.clone(),
-            model_requested: existing.model_requested.clone(),
-            model_served: existing.model_served.clone(),
-            gateway_host: existing.gateway_host.clone(),
-        },
-    )?;
-    if let Some(path) = existing.exported_path.as_deref() {
-        overwrite_exported_note_guarded(state, &meeting_id, &existing.provider_id, path, &enriched)?;
-    }
-    Ok(NoteDto {
-        meeting_id,
-        provider_id: existing.provider_id,
-        markdown: enriched,
-        exported_path: existing.exported_path,
-    })
-}
-
 /// Max cross-meeting links Lane A appends to a note (small + high-precision — a note gains a handful
 /// of related links, not a research dump).
 const MAX_RELATED_LINKS: usize = 4;
@@ -2572,834 +2307,6 @@ fn index_note_body_chunks(
         .index_note_chunks(id, title, &body, embedder)
 }
 
-// ── NOTES — selection Brain-assistant (WP4) ──────────────────────────────────────────────────────
-//
-// The editor's selection popover calls `note_assistant_action`. Refine/Shorten rewrite the
-// selection; Enhance retrieves related brain context (VISIBLE sources only, excluding the current
-// note) and proposes an ADDITIVE passage with citations. Routing is `provider_for(Role::Notes)` —
-// which gives local-Qwen-vs-cloud-Claude selection, the fail-closed consent gate, the
-// `RedactingProvider` firewall, and the egress ledger FOR FREE — never a direct provider build.
-
-/// RAII guard for ONE note-assist turn (residual W7 — the command-surface twin of
-/// `transcribe::live::TurnGuard`): on drop — normal return, gate refusal, provider error, or a
-/// panic unwinding through the async body — it decrements the per-note in-flight counter and,
-/// when THIS turn raised it (`priority`, set only for a LOCAL decode), clears the user-turn
-/// priority flag. The conditional clear means a concurrent live turn's flag is never stomped.
-struct NoteAssistTurnGuard<'a> {
-    state: &'a AppState,
-    key: String,
-    priority: bool,
-}
-impl Drop for NoteAssistTurnGuard<'_> {
-    fn drop(&mut self) {
-        if self.priority {
-            self.state
-                .user_turn_in_progress
-                .store(false, std::sync::atomic::Ordering::Relaxed);
-        }
-        crate::transcribe::live::end_turn(&self.state.in_flight_turns, &self.key);
-    }
-}
-
-/// The FULL known note-assistant action set (the FE catalog mirror). `custom` is always available
-/// (the escape hatch) so it is intentionally OUTSIDE this list — it is handled explicitly. An action
-/// not in this list and not `custom` is an unknown id → `InvalidArg`.
-const NOTE_ASSIST_KNOWN_ACTIONS: &[&str] = &[
-    // EDIT (replace)
-    "refine",
-    "grammar",
-    "shorten",
-    "expand",
-    "simplify",
-    "tone",
-    "translate",
-    // STRUCTURE
-    "bullets",
-    "table",
-    "keypoints",
-    // FROM YOUR BRAIN
-    "enhance",
-    "find_related",
-    "link_entities",
-    "fact_check",
-    "ask",
-    // EXTRACT
-    "action_items",
-    "decisions",
-    // CREATE (artifact)
-    "draft_followup",
-    "spinoff_note",
-];
-
-/// The result shape the FE renders + applies off (see the seam contract table). MUST be one of
-/// `"replace" | "insert" | "info" | "artifact"`. `custom` is a free-text replace.
-fn note_assist_shape(action: &str) -> &'static str {
-    match action {
-        // EDIT + link_entities + custom rewrite the selection in place.
-        "refine" | "grammar" | "shorten" | "expand" | "simplify" | "tone" | "translate"
-        | "bullets" | "table" | "link_entities" | "custom" => "replace",
-        // Keeps the text; appends after the selection.
-        "keypoints" | "enhance" | "action_items" | "decisions" => "insert",
-        // Read-only grounded answer + citations (no destructive edit).
-        "find_related" | "fact_check" | "ask" => "info",
-        // A drafted email/note (title + body).
-        "draft_followup" | "spinoff_note" => "artifact",
-        // Unknown ids never reach here (gated to InvalidArg upstream); default to the safest
-        // non-destructive shape.
-        _ => "info",
-    }
-}
-
-/// Which citation-gathering strategy an action uses. Grounded brain actions reuse the enhance
-/// readers (visibility-gated); `link_entities` uses the gated entity list; everything else needs
-/// no retrieval.
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum NoteAssistRetrieval {
-    /// No retrieval (pure edit on the selection).
-    None,
-    /// The enhance readers: `search_visible` + `search_doc_chunks_*_visible` (excluding this note).
-    BrainCitations,
-    /// The gated entity list (`list_entities_visible`) → which names to wikilink.
-    Entities,
-}
-
-fn note_assist_retrieval(action: &str) -> NoteAssistRetrieval {
-    match action {
-        "enhance" | "find_related" | "fact_check" | "ask" => NoteAssistRetrieval::BrainCitations,
-        "link_entities" => NoteAssistRetrieval::Entities,
-        _ => NoteAssistRetrieval::None,
-    }
-}
-
-/// The selection Brain-assistant action. GATED five ways (normative order): (1) the action must be
-/// ENABLED in config (else `Unavailable`); (2) the note's folder must be unlocked (never send a
-/// sealed note's text off-device / to any model, else `Locked`); (3) brain-grounded retrieval
-/// contributes ONLY visible/unlocked sources; (4) the cloud path rides the redaction firewall via
-/// `provider_for`. `find_related` is retrieval-ONLY (no provider, no egress). Returns the suggestion
-/// + citations + display metadata (modelLabel/mode/redacted/shape/title).
-#[tauri::command]
-pub async fn note_assistant_action(
-    state: State<'_, AppState>,
-    req: NoteAssistRequest,
-) -> Result<NoteAssistResult, AppError> {
-    note_assistant_action_inner(state.inner(), req).await
-}
-
-/// Core of [`note_assistant_action`] over `&AppState` (unit-testable headless). The gate order is
-/// normative: config-enabled → note-unlocked → build provider (consent/firewall) → retrieve → call.
-pub(crate) async fn note_assistant_action_inner(
-    state: &AppState,
-    req: NoteAssistRequest,
-) -> Result<NoteAssistResult, AppError> {
-    // Production path: build the NOTES-role provider through the full egress gate chain.
-    note_assistant_action_impl(state, req, None).await
-}
-
-/// The testable core of [`note_assistant_action_inner`]. `provider_override` lets a unit test inject
-/// a scripted fake provider (so the provider-dependent actions run headless without shelling out to
-/// a real LLM); in production it is `None` and the NOTES-role provider is built via `provider_for`
-/// (consent gate + redaction firewall + egress ledger). The gate order is IDENTICAL either way — the
-/// override only replaces the provider CONSTRUCTION at step (6), never a gate.
-async fn note_assistant_action_impl(
-    state: &AppState,
-    req: NoteAssistRequest,
-    provider_override: Option<std::sync::Arc<dyn crate::summarize::provider::SummarizerProvider>>,
-) -> Result<NoteAssistResult, AppError> {
-    let action = req.action.trim().to_lowercase();
-    // (1) ACTION ENABLED? A disabled action is refused BEFORE any read/egress.
-    let config = {
-        state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?
-            .clone()
-    };
-    // The 3 legacy actions keep their own bools (backward compat — the FE still sends all three).
-    // `custom` is the always-on escape hatch. Every OTHER KNOWN action is enabled UNLESS the user
-    // opted it OUT (`note_assist_actions_off`). An id that is neither known nor `custom` → InvalidArg
-    // BEFORE any read/egress.
-    let enabled = match action.as_str() {
-        "refine" => config.note_assist_refine,
-        "shorten" => config.note_assist_shorten,
-        "enhance" => config.note_assist_enhance,
-        "custom" => true,
-        other if NOTE_ASSIST_KNOWN_ACTIONS.contains(&other) => {
-            !config.note_assist_actions_off.iter().any(|a| a == other)
-        }
-        other => {
-            return Err(AppError::InvalidArg(format!(
-                "unknown note-assistant action: {other}"
-            )));
-        }
-    };
-    if !enabled {
-        return Err(AppError::Unavailable(format!(
-            "the {action} note action is turned off in Settings"
-        )));
-    }
-    if req.selection.trim().is_empty() {
-        return Err(AppError::InvalidArg("no text selected".into()));
-    }
-
-    // TURN DISCIPLINE (residual W7 — Brain v2 P0.3 parity with `spawn_assistant_turn`): at most ONE
-    // note-assist turn per note id at a time. A second call while one is in flight is refused (the
-    // double-click pile-up guard), so duplicate decodes never stack generations on shared Metal.
-    // The key is namespaced (`note-assist:<id>`) so it can never collide with the live loop's
-    // meeting-id keys. Opaque id only — no PII.
-    let turn_key = format!("note-assist:{}", req.note_id);
-    if !crate::transcribe::live::try_begin_turn(&state.in_flight_turns, &turn_key) {
-        return Err(AppError::Unavailable(
-            "the note assistant is already working on this note — wait for it to finish".into(),
-        ));
-    }
-    // RAII: released on EVERY exit path below (gate refusal, provider error, success) — a wedged
-    // key can never permanently refuse this note. Mirrors `transcribe::live::TurnGuard`.
-    let mut turn = NoteAssistTurnGuard {
-        state,
-        key: turn_key,
-        priority: false,
-    };
-
-    // (2) READ-GATE: the note's folder must be unlocked (never egress a sealed note's text).
-    let Some(row) = state.db.get_note_row(&req.note_id)? else {
-        return Err(AppError::InvalidArg(format!("no note {}", req.note_id)));
-    };
-    if !folder_is_unlocked(state, &row.folder_id)? {
-        return Err(AppError::Locked(
-            "this note is locked — unlock its folder to use the assistant".into(),
-        ));
-    }
-
-    let shape = note_assist_shape(&action).to_string();
-    let retrieval = note_assist_retrieval(&action);
-
-    // (3) FIND_RELATED: retrieval-ONLY — NO provider, NO egress. Gather visible citations through
-    //     the SAME gated readers as enhance, build a one-line answer, and return `shape="info"`
-    //     with `mode="local"`, `redacted=false`. This never raises the local user-turn priority
-    //     flag (no decode contends for Metal) and never calls a model → a privacy win.
-    if action == "find_related" {
-        let citations = gather_note_enhance_citations(state, &config, &req)?;
-        let suggestion = match citations.len() {
-            0 => "No related sources found in your brain.".to_string(),
-            1 => "1 related source in your brain.".to_string(),
-            n => format!("{n} related sources in your brain."),
-        };
-        tracing::info!(
-            target: "notes",
-            action = %action,
-            mode = "local",
-            citations = citations.len(),
-            redacted = false,
-            "note assistant action completed (retrieval-only)"
-        );
-        return Ok(NoteAssistResult {
-            action,
-            suggestion,
-            citations,
-            model_label: "Your brain (local search)".to_string(),
-            mode: "local".to_string(),
-            redacted: false,
-            shape,
-            title: None,
-        });
-    }
-
-    // (4) Resolve the display metadata (modelLabel/mode) from the RESOLVED target BEFORE the call.
-    let target =
-        crate::summarize::roles::provider_target(crate::summarize::roles::Role::Notes, &config);
-    let mode = if crate::summarize::egress_is_cloud(&target.connection, &config) {
-        "cloud"
-    } else {
-        "local"
-    };
-    // USER-TURN PRIORITY (residual W7): a LOCAL note-assist decode contends for the on-device
-    // engine (shared Metal) — raise the priority flag for the turn's duration so the background
-    // Realtime-Reactions scan defers, exactly like the live loop's assistant turns. The guard
-    // clears it on every exit path; a cloud call never touches the flag.
-    if mode == "local" {
-        state
-            .user_turn_in_progress
-            .store(true, std::sync::atomic::Ordering::Relaxed);
-        turn.priority = true;
-    }
-    let model_requested =
-        crate::summarize::effective_model_requested(&target, &config);
-    let conn_label = crate::summarize::roles::connection_display_name(&target.connection);
-    let model_label = if model_requested.trim().is_empty() {
-        conn_label.to_string()
-    } else {
-        format!("{conn_label} · {model_requested}")
-    };
-
-    // (5) Retrieve VISIBLE grounding for the brain-grounded actions (EXCLUDING this note). Both the
-    //     citation readers and `list_entities_visible` push the session unlock set through
-    //     `visibility_clause`, so a sealed source never grounds a result.
-    let (citations, entity_names) = match retrieval {
-        NoteAssistRetrieval::BrainCitations => {
-            (gather_note_enhance_citations(state, &config, &req)?, Vec::new())
-        }
-        NoteAssistRetrieval::Entities => {
-            let unlocked = unlocked_snapshot(state)?;
-            let names: Vec<String> = state
-                .db
-                .list_entities_visible(&unlocked)?
-                .into_iter()
-                .map(|n| n.name)
-                .collect();
-            (Vec::new(), names)
-        }
-        NoteAssistRetrieval::None => (Vec::new(), Vec::new()),
-    };
-
-    // (6) Build the prompts, then call the NOTES-role provider (consent gate + redaction firewall +
-    //     egress ledger ride inside `provider_for`/`complete_with_meta_opts`). The edit runs under a
-    //     per-action token cap + low temperature (`GenOptions::edit_rewrite`) so a compression edit
-    //     can't run away and LENGTHEN, and `generate_note_edit` enforces "shorten is actually shorter"
-    //     with one stricter retry.
-    let (system, user) =
-        build_note_assist_prompt(&action, &req, &citations, &entity_names, &config.note_language);
-    let provider = match provider_override {
-        Some(p) => p,
-        None => crate::summarize::provider_for(
-            crate::summarize::roles::Role::Notes,
-            &config,
-            &state.heavy_inference,
-        )?,
-    };
-    let opts = crate::reason::GenOptions::edit_rewrite(note_edit_max_tokens(
-        &action,
-        req.selection.chars().count(),
-    ));
-    let input_words = note_edit_word_count(&req.selection);
-    let (mut suggestion, meta) =
-        generate_note_edit(provider.as_ref(), &action, &system, &user, opts, input_words).await?;
-    suggestion = suggestion.trim().to_string();
-
-    // Artifacts carry a title (email subject / note title). Derive it from the note's own title for
-    // a follow-up draft, and from the drafted body's first line for a spin-off note (a title only —
-    // never logged). Non-artifacts have no title.
-    let title = if shape == "artifact" {
-        Some(derive_artifact_title(
-            &action,
-            row.title.as_deref().unwrap_or(""),
-            &suggestion,
-        ))
-    } else {
-        None
-    };
-
-    // `redacted` = the firewall scrubbed at least one PII token on THIS call (only a cloud
-    // RedactingProvider populates `meta.redactions`; a local provider leaves it None → false).
-    let redacted = meta
-        .redactions
-        .as_ref()
-        .map(|r| r.email + r.card + r.phone + r.name > 0)
-        .unwrap_or(false);
-
-    tracing::info!(
-        target: "notes",
-        action = %action,
-        mode = %mode,
-        shape = %shape,
-        citations = citations.len(),
-        redacted,
-        "note assistant action completed"
-    );
-
-    Ok(NoteAssistResult {
-        action,
-        suggestion,
-        citations,
-        model_label,
-        mode: mode.to_string(),
-        redacted,
-        shape,
-        title,
-    })
-}
-
-/// Derive a non-PII-in-logs artifact title. `draft_followup` reuses the note's own title as an email
-/// subject; `spinoff_note` uses the drafted body's first non-empty line (stripped of leading `#`),
-/// falling back to the note title. The returned string is user-facing content (a title) — it is NEVER
-/// logged.
-fn derive_artifact_title(action: &str, note_title: &str, body: &str) -> String {
-    let fallback = |t: &str| {
-        let t = t.trim();
-        if t.is_empty() {
-            "Untitled".to_string()
-        } else {
-            t.to_string()
-        }
-    };
-    match action {
-        "draft_followup" => {
-            let subj = note_title.trim();
-            if subj.is_empty() {
-                "Follow-up".to_string()
-            } else {
-                format!("Re: {subj}")
-            }
-        }
-        // spinoff_note: first meaningful body line as the title.
-        _ => body
-            .lines()
-            .map(|l| l.trim_start_matches('#').trim())
-            .find(|l| !l.is_empty())
-            .map(|l| l.chars().take(120).collect::<String>())
-            .filter(|l| !l.is_empty())
-            .unwrap_or_else(|| fallback(note_title)),
-    }
-}
-
-/// enhance-context retrieval: run the GATED brain readers (meeting `search_visible` + document/note
-/// `search_doc_chunks_*_visible`), EXCLUDE the current note's own document id, cap at ≤6, and build
-/// [`NoteCitation`]s. Only VISIBLE/unlocked sources contribute (both readers push the live session
-/// unlock set through `visibility_clause`), so a sealed source never grounds an enhancement.
-fn gather_note_enhance_citations(
-    state: &AppState,
-    config: &AppConfig,
-    req: &NoteAssistRequest,
-) -> Result<Vec<NoteCitation>, AppError> {
-    const MAX_CITATIONS: usize = 6;
-    let unlocked = unlocked_snapshot(state)?;
-    // The query is the selection plus a little surrounding context (better recall than the raw
-    // selection alone); the readers tokenize/defuse it safely.
-    let mut query = req.selection.clone();
-    if let Some(b) = &req.before {
-        query.push(' ');
-        query.push_str(b);
-    }
-    if let Some(a) = &req.after {
-        query.push(' ');
-        query.push_str(a);
-    }
-
-    let mut out: Vec<NoteCitation> = Vec::new();
-
-    // Meeting notes/segments (FTS over visible meetings).
-    for hit in state.db.search_visible(&query, MAX_CITATIONS as i64, &unlocked)? {
-        out.push(NoteCitation {
-            kind: "meeting".into(),
-            id: hit.meeting.id.clone(),
-            title: hit.meeting.title.clone().unwrap_or_else(|| "Meeting".into()),
-            snippet: hit.snippet,
-        });
-        if out.len() >= MAX_CITATIONS {
-            break;
-        }
-    }
-
-    // Other notes/documents (semantic when the e5 model is present, else FTS). EXCLUDE the current
-    // note's own document id (never cite the note being edited).
-    if out.len() < MAX_CITATIONS {
-        let doc_hits = if config.semantic_search_enabled && crate::embed::embed_model_present() {
-            let embedder = crate::embed::active_embedder();
-            let qvecs = embedder.embed_query(std::slice::from_ref(&query))?;
-            match qvecs.into_iter().next() {
-                Some(qvec) => state
-                    .db
-                    .search_doc_chunks_visible(&qvec, MAX_CITATIONS as i64, &unlocked)?,
-                None => Vec::new(),
-            }
-        } else {
-            state
-                .db
-                .search_doc_chunks_fts_visible(&query, MAX_CITATIONS as i64, &unlocked)?
-        };
-        for hit in doc_hits {
-            if hit.document_id == req.note_id {
-                continue; // never cite the note being edited.
-            }
-            out.push(NoteCitation {
-                kind: "note".into(),
-                id: hit.document_id,
-                title: hit.name,
-                snippet: hit.snippet,
-            });
-            if out.len() >= MAX_CITATIONS {
-                break;
-            }
-        }
-    }
-
-    // Org shared brain (deliberately-disclosed colleague content — outside the folder-lock domain,
-    // gated on membership only via `org_brain_available`, same seam as the `org_brain_search` agent
-    // tool / MCP `org_search`). RETRIEVAL-ONLY: no provider call, no egress — this is a private,
-    // user-navigated discovery surface, matching `find_related`'s zero-provider-call invariant.
-    if out.len() < MAX_CITATIONS && crate::tools::org_brain_available(&state.db, config) {
-        let org_hits = crate::tools::search_org_brain_hits(&state.db, config, &query)?;
-        for hit in org_hits {
-            out.push(NoteCitation {
-                kind: "org".into(),
-                id: hit.item_id,
-                title: hit.title,
-                snippet: hit.snippet,
-            });
-            if out.len() >= MAX_CITATIONS {
-                break;
-            }
-        }
-    }
-    Ok(out)
-}
-
-/// Format the retrieved brain citations as a numbered grounding block for the grounded actions.
-fn note_assist_grounding_block(citations: &[NoteCitation]) -> String {
-    let mut grounding = String::new();
-    for (i, c) in citations.iter().enumerate() {
-        grounding.push_str(&format!(
-            "[{n}] ({kind}) {title}: {snippet}\n",
-            n = i + 1,
-            kind = c.kind,
-            title = c.title,
-            snippet = c.snippet
-        ));
-    }
-    if grounding.is_empty() {
-        grounding.push_str("(no related material found)\n");
-    }
-    grounding
-}
-
-/// Build the (system, user) prompts for a note-assistant action. EDIT actions rewrite the selection
-/// with its surrounding context; STRUCTURE reshapes it; the FROM-YOUR-BRAIN actions ground on the
-/// retrieved citations (or entity names) ONLY; the INFO actions (`fact_check`/`ask`) produce an
-/// ANSWER, not an edit; CREATE actions draft an artifact body. `note_language` steers the reply
-/// language (matching the rest of the note stack). Every prompt passes the preceding text as
-/// READ-ONLY context ("do NOT continue it") with the SELECTION LAST — the discipline that fixed
-/// "shorten made it longer".
-fn build_note_assist_prompt(
-    action: &str,
-    req: &NoteAssistRequest,
-    citations: &[NoteCitation],
-    entity_names: &[String],
-    note_language: &str,
-) -> (String, String) {
-    let lang = if note_language.trim().is_empty() || note_language == "auto" {
-        "the same language as the selected text".to_string()
-    } else {
-        format!("language code '{note_language}'")
-    };
-    // EDIT actions (refine/grammar/shorten/expand/simplify/tone) rewrite a passage the user
-    // ALREADY wrote in some language — they must always match ITS language, never the global
-    // `note_language` pin (that pin is for content GENERATED from scratch: the full note, the
-    // STRUCTURE/FROM-YOUR-BRAIN/INFO/CREATE actions, and `translate`'s explicit target). Forcibly
-    // translating during a surgical edit was the bug: Shorten on an English passage under a
-    // Polish-pinned note_language rewrote it into Polish instead of shortening it in English.
-    let edit_lang = "the same language as the selected text".to_string();
-    let before = req.before.as_deref().unwrap_or("");
-    let preceding = if before.trim().is_empty() {
-        String::new()
-    } else {
-        format!("Preceding text (READ-ONLY context — do NOT reproduce or continue it):\n{before}\n\n")
-    };
-    let sel = req.selection.as_str();
-    let variant = req.variant.as_deref().unwrap_or("").trim();
-    let instruction = req.instruction.as_deref().unwrap_or("").trim();
-    match action {
-        "refine" => {
-            let system = format!(
-                "You refine a passage of the user's own note: improve clarity, grammar, and flow \
-                 WITHOUT changing its meaning, adding facts, or padding its length. Reply in \
-                 {edit_lang}. Output ONLY the rewritten passage — no preamble, no quotes, no \
-                 explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO REFINE (rewrite ONLY this):\n{sel}");
-            (system, user)
-        }
-        "grammar" => {
-            let system = format!(
-                "You are a copy-editor. Correct ONLY spelling, grammar, and punctuation in the \
-                 passage. Do NOT restructure, rephrase, shorten, lengthen, or change the meaning or \
-                 word choice beyond what a correction requires. Reply in {edit_lang}. Output ONLY \
-                 the corrected passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO CORRECT (fix ONLY this):\n{sel}");
-            (system, user)
-        }
-        "shorten" => {
-            let system = format!(
-                "You shorten a passage of the user's own note. Rewrite it in ABOUT HALF the \
-                 sentences: keep every decision, number, name, date, and commitment; cut hedging, \
-                 repetition, filler, and throat-clearing. The result MUST be shorter than the \
-                 original. Reply in {edit_lang}. Output ONLY the shortened passage — no preamble, \
-                 no quotes, no explanation.\n\nExample —\nOriginal: I think that, honestly, we should \
-                 probably consider maybe moving the deadline to Friday, because the team is quite \
-                 busy right now and there is a lot going on.\nShortened: Move the deadline to \
-                 Friday — the team is overloaded."
-            );
-            let user = format!("{preceding}PASSAGE TO SHORTEN (rewrite ONLY this, shorter):\n{sel}");
-            (system, user)
-        }
-        "expand" => {
-            let system = format!(
-                "You expand a terse passage of the user's own note into fuller, clearer prose. \
-                 Elaborate ONLY on what is already stated — spell out shorthand, join fragments into \
-                 sentences, add connective phrasing. Do NOT invent facts, opinions, numbers, names, \
-                 or commitments that are not in the passage. Reply in {edit_lang}. Output ONLY the \
-                 expanded passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO EXPAND (rewrite ONLY this, fuller):\n{sel}");
-            (system, user)
-        }
-        "simplify" => {
-            let system = format!(
-                "You rewrite a passage of the user's own note in plain, jargon-free language a \
-                 non-expert can follow. Keep every fact, number, name, and decision; replace \
-                 jargon and convoluted phrasing with simple words and short sentences. Do NOT add or \
-                 remove information. Reply in {edit_lang}. Output ONLY the simplified passage — no \
-                 preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO SIMPLIFY (rewrite ONLY this):\n{sel}");
-            (system, user)
-        }
-        "tone" => {
-            let tone = if variant.is_empty() { "professional" } else { variant };
-            let system = format!(
-                "You rewrite a passage of the user's own note in a {tone} tone WITHOUT changing its \
-                 meaning, facts, or length beyond what the tone requires. Reply in {edit_lang}. \
-                 Output ONLY the rewritten passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO REWRITE in a {tone} tone (rewrite ONLY this):\n{sel}");
-            (system, user)
-        }
-        "translate" => {
-            // For translate the TARGET language is the variant, overriding note_language.
-            let target = if variant.is_empty() {
-                "the language the user most likely wants".to_string()
-            } else {
-                variant.to_string()
-            };
-            let system = format!(
-                "You translate a passage of the user's own note into {target}. Preserve meaning, \
-                 tone, formatting, names, numbers, and any markdown. Do NOT add or omit content. \
-                 Output ONLY the translated passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO TRANSLATE into {target} (translate ONLY this):\n{sel}");
-            (system, user)
-        }
-        "bullets" => {
-            let system = format!(
-                "You reformat a passage of the user's own note into a markdown bullet list. Turn \
-                 each distinct point into its own `- ` line, preserving every fact, number, name, and \
-                 decision; do NOT add, remove, or invent content. Reply in {lang}. Output ONLY the \
-                 markdown list — no preamble, no heading, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO CONVERT to a bullet list (convert ONLY this):\n{sel}");
-            (system, user)
-        }
-        "table" => {
-            let system = format!(
-                "You reformat a passage of the user's own note into a GitHub-flavored markdown table \
-                 with a header row and a `---` separator. Infer sensible columns from the content; \
-                 preserve every fact, number, name, and decision; do NOT add or invent data. Reply \
-                 in {lang}. Output ONLY the markdown table — no preamble, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO CONVERT to a table (convert ONLY this):\n{sel}");
-            (system, user)
-        }
-        "keypoints" => {
-            let system = format!(
-                "You write a SHORT TL;DR digest of a passage of the user's own note: 2–4 markdown \
-                 bullets capturing only the key points, decisions, and numbers. This is an ADDITIVE \
-                 summary to insert AFTER the selection — do NOT rewrite or reproduce the original. \
-                 Reply in {lang}. Output ONLY the bullet digest — no preamble, no heading, no \
-                 explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO SUMMARIZE (write a short digest of this):\n{sel}");
-            (system, user)
-        }
-        "enhance" => {
-            let grounding = note_assist_grounding_block(citations);
-            let system = format!(
-                "You expand the user's note by proposing a SHORT ADDITIVE passage that builds on \
-                 the selection using ONLY the RELATED MATERIAL provided — never invent facts. If the \
-                 material adds nothing, reply with an empty line. Reply in {lang}. Output ONLY the \
-                 additive passage to INSERT after the selection — no preamble, no headings, no \
-                 explanation."
-            );
-            let user = format!(
-                "RELATED MATERIAL (from the user's own brain):\n{grounding}\nSELECTION TO EXPAND:\n{sel}"
-            );
-            (system, user)
-        }
-        "link_entities" => {
-            let names = if entity_names.is_empty() {
-                "(none — return the selection unchanged)".to_string()
-            } else {
-                entity_names.join(", ")
-            };
-            let system = format!(
-                "You rewrite a passage of the user's own note, wrapping ONLY the known entity names \
-                 listed below in `[[wikilinks]]` where they appear. Do NOT invent links, do NOT link \
-                 any name not in the list, do NOT change any other word, spacing, or punctuation, and \
-                 do NOT double-wrap a name already inside `[[...]]`. If a name does not appear in the \
-                 passage, leave the passage as-is for that name. Reply in {lang}. Output ONLY the \
-                 rewritten passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!(
-                "KNOWN ENTITY NAMES (link ONLY these):\n{names}\n\n{preceding}PASSAGE TO LINK (rewrite ONLY this):\n{sel}"
-            );
-            (system, user)
-        }
-        "fact_check" => {
-            let grounding = note_assist_grounding_block(citations);
-            let system = format!(
-                "You fact-check the SELECTION against the user's OWN brain (the RELATED MATERIAL \
-                 below) ONLY — never external knowledge. Flag any claim in the selection that \
-                 CONTRADICTS or is UNSUPPORTED by the material, quoting the conflicting source. If \
-                 everything checks out, say so briefly. This is an ANSWER, not an edit — do NOT \
-                 rewrite the selection. Reply in {lang}. Output ONLY your findings — no preamble."
-            );
-            let user = format!(
-                "RELATED MATERIAL (from the user's own brain):\n{grounding}\nSELECTION TO FACT-CHECK:\n{sel}"
-            );
-            (system, user)
-        }
-        "ask" => {
-            let grounding = note_assist_grounding_block(citations);
-            let question = if instruction.is_empty() {
-                "What is the most important thing to know about this selection?"
-            } else {
-                instruction
-            };
-            let system = format!(
-                "You answer the user's QUESTION about the SELECTION, grounded in the SELECTION and \
-                 the RELATED MATERIAL from their own brain ONLY — never invent facts. If the answer \
-                 is not in the material, say so. This is an ANSWER, not an edit — do NOT rewrite the \
-                 selection. Reply in {lang}. Output ONLY the answer — no preamble."
-            );
-            let user = format!(
-                "QUESTION:\n{question}\n\nRELATED MATERIAL (from the user's own brain):\n{grounding}\nSELECTION:\n{sel}"
-            );
-            (system, user)
-        }
-        "action_items" => {
-            let system = format!(
-                "You extract action items / TODOs from a passage of the user's own note into a \
-                 markdown checklist. Each task is its own `- [ ] ` line; capture the owner and any \
-                 due date if stated; do NOT invent tasks. If there are no action items, reply with an \
-                 empty line. This is an ADDITIVE list to insert AFTER the selection — do NOT reproduce \
-                 the original. Reply in {lang}. Output ONLY the checklist — no preamble, no heading."
-            );
-            let user = format!("{preceding}PASSAGE TO SCAN for action items:\n{sel}");
-            (system, user)
-        }
-        "decisions" => {
-            let system = format!(
-                "You extract the DECISIONS made in a passage of the user's own note into a short \
-                 markdown bullet list. Capture only decisions actually made (not open questions or \
-                 tasks); do NOT invent any. If there are no decisions, reply with an empty line. This \
-                 is an ADDITIVE list to insert AFTER the selection — do NOT reproduce the original. \
-                 Reply in {lang}. Output ONLY the list — no preamble, no heading."
-            );
-            let user = format!("{preceding}PASSAGE TO SCAN for decisions:\n{sel}");
-            (system, user)
-        }
-        "draft_followup" => {
-            let system = format!(
-                "You draft a concise follow-up email or message based on the SELECTION from the \
-                 user's own note. Cover the key points, decisions, and next steps; keep a {tone} \
-                 tone. Do NOT invent facts beyond the selection. Do NOT include a subject line — just \
-                 the message body. Reply in {lang}. Output ONLY the message body — no preamble, no \
-                 explanation.",
-                tone = if variant.is_empty() { "professional" } else { variant }
-            );
-            let user = format!("{preceding}SELECTION TO TURN INTO A FOLLOW-UP MESSAGE:\n{sel}");
-            (system, user)
-        }
-        "custom" => {
-            // The free-text "Ask Brain to edit…" instruction, applied to the SELECTION (shape=replace).
-            // MUST have its own arm — without it `custom` fell through to the spinoff_note catch-all,
-            // silently dropping the instruction and drafting an unrelated note that Accept then
-            // destructively wrote over the selection. The instruction is woven into the directive; the
-            // FE only sends `custom` with non-empty text, but an empty directive degrades to a refine.
-            let directive = if instruction.is_empty() {
-                "Improve the clarity, grammar, and flow of the passage without changing its meaning"
-            } else {
-                instruction
-            };
-            let system = format!(
-                "You edit a passage of the user's own note by applying THIS instruction to it: \
-                 \"{directive}\". Apply the instruction to the passage ONLY — do NOT invent facts \
-                 beyond it, do NOT answer as chat, do NOT continue the surrounding text. Reply in \
-                 {lang}. Output ONLY the edited passage — no preamble, no quotes, no explanation."
-            );
-            let user = format!("{preceding}PASSAGE TO EDIT (apply the instruction, rewrite ONLY this):\n{sel}");
-            (system, user)
-        }
-        // spinoff_note
-        _ => {
-            let system = format!(
-                "You draft a new standalone note from the SELECTION in the user's existing note. \
-                 Start with a short `# ` heading that titles the new note, then write clean note body \
-                 in markdown. Build ONLY on the selection — do NOT invent facts. Reply in {lang}. \
-                 Output ONLY the new note (heading + body) — no preamble, no explanation."
-            );
-            let user = format!("{preceding}SELECTION TO TURN INTO A NEW NOTE:\n{sel}");
-            (system, user)
-        }
-    }
-}
-
-/// Whitespace word count — the unit the note-edit length guard compares in (word/sentence targets
-/// are followable by small models; character/token targets are not).
-fn note_edit_word_count(s: &str) -> usize {
-    s.split_whitespace().count()
-}
-
-/// The RUNAWAY-GUARD token cap for a note edit, sized off the selection (rough chars/4 token
-/// estimate). `shorten` is capped at ~the input length so it can't physically LENGTHEN; the
-/// in-place edits (refine/grammar/simplify/tone/translate/bullets/table/link_entities/custom) get
-/// modest headroom; the ADDITIVE / GENERATIVE actions (expand/enhance/keypoints/action_items/
-/// decisions/fact_check/ask/drafts) get GENEROUS headroom because their output is new content, not a
-/// rewrite of the selection. Floors keep a tiny selection from being truncated; ceilings keep a huge
-/// selection inside the 4096-token on-device context budget. A safety net — the prompt's length
-/// budget + [`generate_note_edit`]'s validation are the primary length controls.
-fn note_edit_max_tokens(action: &str, input_chars: usize) -> usize {
-    let input_tokens = (input_chars / 4).max(1);
-    let (mult, floor, ceil) = match action {
-        // Must not exceed ~input tokens (physically can't lengthen).
-        "shorten" => (1.0_f64, 48usize, 1024usize),
-        // Additive / generated output: the answer/draft/list is NEW content — give it room, and a
-        // higher floor so a short selection can still yield a full draft or answer.
-        "expand" | "enhance" | "keypoints" | "action_items" | "decisions" | "fact_check" | "ask"
-        | "draft_followup" | "spinoff_note" => (3.0_f64, 256usize, 2048usize),
-        // In-place edits can legitimately match or slightly exceed the input length.
-        _ => (1.5_f64, 64usize, 1536usize),
-    };
-    ((input_tokens as f64 * mult).ceil() as usize).clamp(floor, ceil)
-}
-
-/// Generate one note edit through `provider`, and for `shorten` ENFORCE that the result is actually
-/// shorter than the input — one stricter retry if the first attempt is not (the "shorten made it
-/// longer" guard; the model otherwise ran unbounded on the fully-local path). Returns the shortest
-/// candidate for `shorten`, the single result otherwise. Takes `&dyn SummarizerProvider` so it is
-/// unit-testable with a scripted fake provider.
-async fn generate_note_edit(
-    provider: &dyn crate::summarize::provider::SummarizerProvider,
-    action: &str,
-    system: &str,
-    user: &str,
-    opts: crate::reason::GenOptions,
-    input_words: usize,
-) -> Result<(String, crate::summarize::meta::CallMeta), AppError> {
-    // `opts` is `Copy` — each call gets its own copy (no `.clone()`, which would trip clippy).
-    let (out, meta) = provider.complete_with_meta_opts(system, user, opts).await?;
-    if action == "shorten" && note_edit_word_count(&out) >= input_words {
-        // First attempt did not shorten — retry ONCE with a stricter instruction.
-        let strict = format!(
-            "{system}\n\nThe previous attempt was NOT shorter than the original. Return a STRICTLY \
-             shorter version: fewer words than the original, keeping only the essential facts."
-        );
-        let (out2, meta2) = provider.complete_with_meta_opts(&strict, user, opts).await?;
-        if note_edit_word_count(&out2) < note_edit_word_count(&out) {
-            return Ok((out2, meta2));
-        }
-    }
-    Ok((out, meta))
-}
-
 // ── NOTES — auto-organize (WP5) ──────────────────────────────────────────────────────────────────
 //
 // Two-step, non-destructive. `plan_organize_notes` PROPOSES a target note-folder per visible note
@@ -3537,96 +2444,6 @@ pub(crate) fn apply_organize_plan_inner(
 // SOURCE meeting is sealed-and-not-session-unlocked is INVISIBLE here (and injected into no prompt),
 // because `list_user_facts_visible` filters by source-meeting `visibility_clause` on the live
 // unlocked snapshot. Forget/clear are bitemporal INVALIDATE (close valid_to), never a silent delete.
-
-/// Brain v2 L2.3 — IMPORT pasted memories from another AI assistant (a ChatGPT/Claude "what I
-/// remember about you" export) into the user-memory store. Returns the number of NEW facts added.
-///
-/// Flow: extract candidates on the ON-DEVICE light reasoner (`user_memory::extract_imported_memories`
-/// — stub ⇒ empty ⇒ 0) → deterministic `reconcile_facts` against ALL existing user facts (a
-/// re-import of the same text reconciles to NoOps ⇒ 0 new) → ONLY when there is ≥1 Add, create a
-/// SYNTHETIC anchor meeting (`import-<uuid>`, title "Memory Import", `Exported`, no audio, no
-/// folder — so its facts are VISIBLE via the no-note arm of the visibility predicate) → stamp +
-/// apply atomically. ORDER IS LOAD-BEARING: the meeting row is created only after reconcile found
-/// something to add, so a stub/duplicate import leaves NO synthetic meeting behind. Deleting that
-/// meeting undoes the whole import (`delete_meeting` purges its `user_facts` in-tx). ZERO egress:
-/// extraction runs on [`import_extraction_reasoner`] — LOCAL-or-stub, NEVER cloud (the FE copy
-/// promises on-device; a pasted third-party memory export must not ride the cloud Notes provider).
-/// No local model ⇒ 0 imported (the FE hints the model may be missing). Runs on a blocking worker
-/// (a local-model extraction can take seconds). Logs counts only.
-#[tauri::command]
-pub async fn import_memories(state: State<'_, AppState>, text: String) -> Result<usize, AppError> {
-    let db = state.db.clone();
-    let reasoner = import_extraction_reasoner(state.inner());
-    let enabled = user_memory_enabled(state.inner());
-    tauri::async_runtime::spawn_blocking(move || {
-        import_memories_inner(&db, reasoner.as_ref(), enabled, &text)
-    })
-    .await
-    .map_err(|e| AppError::Other(anyhow::anyhow!("import task join failed: {e}")))?
-}
-
-
-/// Inner of [`import_memories`] (unit-testable: Db + reasoner + flag injected).
-pub(crate) fn import_memories_inner(
-    db: &crate::storage::Db,
-    reasoner: &dyn crate::reason::LocalReasoner,
-    memory_enabled: bool,
-    text: &str,
-) -> Result<usize, AppError> {
-    if !memory_enabled {
-        return Err(AppError::InvalidArg(
-            "cross-meeting memory is turned off — enable it in Settings to import memories".into(),
-        ));
-    }
-    if text.trim().is_empty() {
-        return Err(AppError::InvalidArg("nothing to import — paste the memory text".into()));
-    }
-    // 1) Best-effort extraction (stub / decode failure ⇒ empty ⇒ 0 imported, nothing persisted).
-    let candidates = crate::user_memory::extract_imported_memories(reasoner, text);
-    if candidates.is_empty() {
-        return Ok(0);
-    }
-    // 2) Deterministic reconcile against ALL existing user facts — the dedup: re-importing the same
-    //    export yields NoOps only.
-    let existing = db.user_facts_all()?;
-    let at = chrono::Utc::now().to_rfc3339();
-    let mut ops = crate::facts::reconcile_facts(&existing, &candidates, &at);
-    let adds = ops
-        .iter()
-        .filter(|o| matches!(o, crate::facts::FactOp::Add(_)))
-        .count();
-    if adds == 0 {
-        return Ok(0); // pure dedup/no-op import — no synthetic meeting is created.
-    }
-    // 3) The synthetic anchor meeting — created ONLY now that something will be added, so deleting
-    //    it undoes the import and a no-op import leaves nothing behind.
-    let meeting_id = format!("import-{}", uuid::Uuid::new_v4());
-    db.insert_meeting(&crate::storage::models::Meeting {
-        id: meeting_id.clone(),
-        started_at: at.clone(),
-        ended_at: None,
-        title: Some("Memory Import".to_string()),
-        duration_s: 0,
-        audio_path: None,
-        status: crate::storage::models::MeetingStatus::Exported,
-        folder_id: None,
-    })?;
-    // 4) Stamp the anchor onto the Adds (gating + purge anchor) and apply atomically. MEM-1: use the
-    //    import-aware apply so every pre-existing fact this import SUPERSEDES (an Invalidate on a fact
-    //    anchored to another meeting) is linked to the synthetic import id — deleting the import then
-    //    REOPENS those facts, making "delete to undo" a FULL reversal instead of a partial one that
-    //    leaves prior memories permanently closed.
-    crate::facts::set_meeting_id(&mut ops, &meeting_id);
-    db.apply_user_fact_ops_recording_import_supersedes(&ops, &meeting_id)?;
-    tracing::info!(
-        target: "user_memory",
-        meeting_id = %meeting_id,
-        added = adds,
-        "memories imported (anchored to a synthetic meeting)"
-    );
-    Ok(adds)
-}
-
 
 /// Permanently delete a meeting: its audio file, its exported vault note, and all DB rows
 /// (segments, notes, timeline cascade via FK). Irreversible.
@@ -3928,60 +2745,6 @@ pub fn list_open_commitments(
     state.db.list_open_commitments(&unlocked, owner)
 }
 
-/// Rewrite the note's action items into Obsidian Tasks format (📅 due dates) + re-write the
-/// vault file in place. Returns the updated note.
-#[tauri::command]
-pub fn patch_note_tasks(
-    state: State<'_, AppState>,
-    meeting_id: String,
-) -> Result<NoteDto, AppError> {
-    // BLK-1 / TOCTOU (2026-07-10 audit F4): hold the lifecycle guard across gate+write so a
-    // concurrent relock/seal cannot land between the unlock check and the upsert.
-    let _lifecycle = lifecycle_guard(state.inner());
-    // D4 WRITE-GATE: refuse to rewrite a sealed-and-not-unlocked meeting's note (its plaintext is
-    // blanked; patching would persist the blanked value over the sealed content). Fail closed.
-    if !meeting_is_unlocked(state.inner(), &meeting_id)? {
-        return Err(AppError::Locked(
-            "this meeting's folder is locked — unlock it to rewrite the note's tasks".into(),
-        ));
-    }
-    let existing = state
-        .db
-        .get_latest_note_for_meeting(&meeting_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no note for meeting {meeting_id}")))?;
-    let patched = crate::summarize::action_items::patch_tasks_markdown(&existing.markdown);
-    let created_at = chrono::Utc::now().to_rfc3339();
-    // Seal-on-write (audit F1): a session-unlocked LOCKED folder re-seals the patched markdown.
-    upsert_note_reseal_if_locked(
-        state.inner(),
-        &NoteRecord {
-            meeting_id: meeting_id.clone(),
-            provider_id: existing.provider_id.clone(),
-            markdown: patched.clone(),
-            created_at,
-            exported_path: existing.exported_path.clone(),
-            model_requested: existing.model_requested.clone(),
-            model_served: existing.model_served.clone(),
-            gateway_host: existing.gateway_host.clone(),
-        },
-    )?;
-    if let Some(path) = existing.exported_path.as_deref() {
-        overwrite_exported_note_guarded(
-            state.inner(),
-            &meeting_id,
-            &existing.provider_id,
-            path,
-            &patched,
-        )?;
-    }
-    Ok(NoteDto {
-        meeting_id,
-        provider_id: existing.provider_id,
-        markdown: patched,
-        exported_path: existing.exported_path,
-    })
-}
-
 /// Pin a meeting moment: append a timestamped ^block-ref to the note (DB + vault file) and
 /// return an obsidian:// deep link to the note.
 #[tauri::command]
@@ -4210,36 +2973,14 @@ pub async fn build_and_persist_entities(
 }
 
 // ── Re-Truth (the vault heals itself) — supersession review + one-tap stamp ──────────────────────
-
-/// One supersession surfaced for review (camelCase for the FE). `sourceNotePath` is the absolute
-/// on-disk `.md` the FE never shows (it shows `sourceNoteTitle`); `applied` reflects whether the
-/// row has already been stamped (always `false` from `preview_supersessions`, which returns the
-/// pending set).
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SupersessionDto {
-    pub id: String,
-    pub entity: String,
-    pub predicate: String,
-    pub old_value: String,
-    pub new_value: String,
-    pub source_note_title: String,
-    pub source_note_path: String,
-    pub source_meeting_id: String,
-    pub superseding_meeting_id: String,
-    /// The superseding note's title — `None` when that note is sealed (never leak a locked title).
-    pub superseding_note_title: Option<String>,
-    pub applied: bool,
-}
-
-/// Result of `apply_supersessions`: how many were stamped vs skipped because their source note sealed
-/// (or lost its vault file) between preview and apply (the prune↔seal TOCTOU discipline).
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ApplyResult {
-    pub applied: usize,
-    pub skipped_sealed: usize,
-}
+//
+// The supersession COMMANDS (`preview_supersessions`/`apply_supersessions`/`undo_supersessions`) +
+// their DTOs (`SupersessionDto`/`ApplyResult`) + the enrich-only helpers (`pristine_note_bytes`/
+// `superseding_link_stem`) were extracted verbatim to `commands/enrich.rs` (God-file split, PURE
+// MOVE — every gate/re-gate/undo body byte-identical, only relocated). The three gate helpers below
+// (`source_is_stampable`/`folder_locked_on_disk`/`note_file_for`) STAY here: the sibling
+// `commands/analytics.rs` (the Re-Truth audit-finding path) also reaches them via `super::…`, and
+// the moved commands reach them the same way.
 
 /// A source note is STAMPABLE iff it is BOTH session-unlocked (content-read gate) AND its folder is
 /// NOT locked-on-disk (write-safety gate). A sealed source's `.md` was deleted on seal and its facts
@@ -4280,383 +3021,6 @@ fn note_file_for(state: &AppState, meeting_id: &str) -> Result<Option<(String, S
         .unwrap_or("")
         .to_string();
     Ok(Some((path, stem)))
-}
-
-/// Preview the PENDING supersessions whose superseding meeting is `meeting_id`. GATED: a row is
-/// included ONLY when its SOURCE meeting is stampable (open-on-disk + unlocked) AND has a vault `.md`;
-/// a sealed-or-unexported source contributes NOTHING. The superseding note's title is surfaced only
-/// when that note is itself unlocked. Returns `[]` when there are none.
-#[tauri::command]
-pub fn preview_supersessions(
-    state: State<'_, AppState>,
-    meeting_id: String,
-) -> Result<Vec<SupersessionDto>, AppError> {
-    preview_supersessions_inner(state.inner(), &meeting_id)
-}
-
-pub(crate) fn preview_supersessions_inner(
-    state: &AppState,
-    meeting_id: &str,
-) -> Result<Vec<SupersessionDto>, AppError> {
-    let rows = state.db.unapplied_supersessions_for(meeting_id)?;
-    let mut out = Vec::new();
-    for r in rows {
-        // GATE the SOURCE note (content-read + write-safety). A sealed/not-unlocked source is dropped.
-        if !source_is_stampable(state, &r.source_meeting_id)? {
-            continue;
-        }
-        // GATE the SUPERSEDING side TOO — defense-in-depth. `new_value` is derived from the superseding
-        // meeting's fact, so a sealed-and-not-session-unlocked superseding meeting must surface NOTHING,
-        // even in the brief race window where `lock_folder` has flipped `locked=1` but not yet purged
-        // this row. Purge-on-seal normally removes it; this second-side gate closes the race regardless.
-        if !meeting_is_unlocked(state, &r.superseding_meeting_id)? {
-            continue;
-        }
-        let Some((path, stem)) = note_file_for(state, &r.source_meeting_id)? else {
-            continue; // no vault file → nothing to stamp/show.
-        };
-        // The superseding meeting is now known-unlocked, so its title is safe to surface (`None` only
-        // when it was never exported to the vault).
-        let superseding_note_title =
-            note_file_for(state, &r.superseding_meeting_id)?.map(|(_, s)| s);
-        out.push(SupersessionDto {
-            id: r.id,
-            entity: r.entity,
-            predicate: r.predicate,
-            old_value: r.old_value,
-            new_value: r.new_value,
-            source_note_title: stem,
-            source_note_path: path,
-            source_meeting_id: r.source_meeting_id,
-            superseding_meeting_id: r.superseding_meeting_id,
-            superseding_note_title,
-            applied: r.applied_at.is_some(),
-        });
-    }
-    Ok(out)
-}
-
-/// APPLY the given supersessions: append a `[!superseded]` callout to each SOURCE note (and a mirror
-/// backlink to the superseding note, when it too is open). RE-GATES each row at apply time — a source
-/// that sealed since preview is SKIPPED (never stamped), the prune↔seal TOCTOU discipline. Snapshots
-/// each note's exact bytes into the row's pre-image BEFORE the (append-only) write, so `undo` restores
-/// them byte-identical. Idempotent: an already-applied row is a no-op, and the callout carries a
-/// stable marker so re-stamping never duplicates.
-#[tauri::command]
-pub fn apply_supersessions(
-    state: State<'_, AppState>,
-    ids: Vec<String>,
-) -> Result<ApplyResult, AppError> {
-    apply_supersessions_inner(state.inner(), &ids)
-}
-
-pub(crate) fn apply_supersessions_inner(
-    state: &AppState,
-    ids: &[String],
-) -> Result<ApplyResult, AppError> {
-    // Seal-vs-write TOCTOU (Phase-0 lock-review follow-up): hold the lifecycle guard across the
-    // per-row re-gate + `.md` writes, so a concurrent lock/relock cannot land between
-    // `source_is_stampable` and the append (the same guard `update_note_inner` and every other
-    // vault-writing command already holds).
-    let _lifecycle = lifecycle_guard(state);
-    let mut applied = 0usize;
-    let mut skipped_sealed = 0usize;
-    let date = chrono::Utc::now().format("%Y-%m-%d").to_string();
-
-    // PER-BATCH PRISTINE CACHE, keyed by note FILE path. Multiple rows in ONE heal touch the SAME
-    // note — every row shares the superseding meeting's note, and two facts from one old note share a
-    // SOURCE note too. The undo pre-image for a file MUST be its PRE-BATCH ("pristine") content,
-    // captured the FIRST time this call touches that path — never the mid-batch, already-stamped
-    // bytes a later row would otherwise read. So all rows sharing a note carry IDENTICAL pristine
-    // pre-images and undo (restore-each-file-once) is order-independent + byte-identical for N≥2.
-    let mut pristine: std::collections::HashMap<String, Vec<u8>> = std::collections::HashMap::new();
-
-    // PRE-PASS: seed the cache from any pre-images ALREADY durably stored by an earlier (crashed)
-    // attempt, keyed by note path — so a row that still has to capture its pre-image finds the
-    // pristine bytes here instead of re-reading a sibling-stamped file, regardless of the id order a
-    // retry arrives in (retry-safe AND order-independent).
-    for id in ids {
-        let Some(row) = state.db.get_supersession(id)? else {
-            continue;
-        };
-        if let Some(pre) = &row.source_pre_image {
-            if let Some((path, _)) = note_file_for(state, &row.source_meeting_id)? {
-                pristine.entry(path).or_insert_with(|| pre.clone());
-            }
-        }
-        if let Some(pre) = &row.superseding_pre_image {
-            if let Some((path, _)) = note_file_for(state, &row.superseding_meeting_id)? {
-                pristine.entry(path).or_insert_with(|| pre.clone());
-            }
-        }
-    }
-
-    for id in ids {
-        let Some(row) = state.db.get_supersession(id)? else {
-            continue; // unknown id — nothing to do.
-        };
-        if row.applied_at.is_some() {
-            continue; // already stamped — idempotent no-op.
-        }
-        // TOCTOU re-gate: the source folder may have sealed since preview. A now-sealed/not-unlocked
-        // (or unexported) source is SKIPPED, never stamped.
-        if !source_is_stampable(state, &row.source_meeting_id)? {
-            skipped_sealed += 1;
-            continue;
-        }
-        let Some((source_path, source_stem)) = note_file_for(state, &row.source_meeting_id)? else {
-            skipped_sealed += 1;
-            continue;
-        };
-
-        // The superseding note gets a backlink ONLY when it is itself open-on-disk + unlocked (never
-        // write into or reference a sealed note). Its stem feeds the source callout's `[[…]]` link.
-        let superseding_open = meeting_is_unlocked(state, &row.superseding_meeting_id)?
-            && !folder_locked_on_disk(state, &row.superseding_meeting_id)?;
-        let superseding_file = if superseding_open {
-            note_file_for(state, &row.superseding_meeting_id)?
-        } else {
-            None
-        };
-
-        // Resolve PRISTINE pre-images from the per-path cache (reads + UTF-8-validates each file on its
-        // first batch-touch; every later row sharing the path reuses the SAME pristine bytes).
-        let source_pre = pristine_note_bytes(&mut pristine, &source_path)?;
-        let superseding_pre = match &superseding_file {
-            Some((p, _)) => Some(pristine_note_bytes(&mut pristine, p)?),
-            None => None,
-        };
-
-        // DURABLE-BEFORE-WRITE: persist the pristine pre-images BEFORE any `.md` write, so a crash
-        // between write and mark-applied still leaves a recoverable un-stamped pre-image. `COALESCE`
-        // never clobbers an already-stored pristine backup and, combined with the pristine cache, the
-        // stored bytes are NEVER a re-snapshot of a stamped file.
-        state.db.store_supersession_pre_images(
-            id,
-            Some(&source_pre),
-            superseding_pre.as_deref(),
-        )?;
-
-        // Stamp the SOURCE note: append the callout to the CURRENT on-disk content (idempotent — a
-        // retry over an already-stamped file is a no-op). The undo pre-image is the pristine cache
-        // copy, never these current bytes.
-        let current_source = std::fs::read_to_string(&source_path)
-            .map_err(|e| AppError::Export(format!("read source note failed: {e}")))?;
-        let new_source = crate::export::obsidian::append_supersession_callout(
-            &current_source,
-            &date,
-            &row.predicate,
-            &row.old_value,
-            &row.new_value,
-            superseding_file.as_ref().map(|(_, s)| s.as_str()),
-        );
-        crate::export::obsidian::overwrite_note(std::path::Path::new(&source_path), &new_source)?;
-        // Export-collision guard: this is a read-modify-write APPEND of the CURRENT file (external
-        // edits are preserved in place by construction — no sibling). The baseline is re-stamped
-        // from the final written content ONLY when the pre-append bytes still matched it — see
-        // the helper's laundering rationale.
-        refresh_meeting_note_exported_hash(
-            state,
-            &row.source_meeting_id,
-            &current_source,
-            &new_source,
-        )?;
-
-        // Stamp the SUPERSEDING backlink (its pristine pre-image is now durably stored). Append to its
-        // CURRENT content (idempotent).
-        if let Some((sup_path, _)) = &superseding_file {
-            let current_sup = std::fs::read_to_string(sup_path)
-                .map_err(|e| AppError::Export(format!("read superseding note failed: {e}")))?;
-            let new_sup = crate::export::obsidian::append_supersedes_callout(
-                &current_sup,
-                &date,
-                &row.predicate,
-                &row.old_value,
-                &row.new_value,
-                &source_stem,
-            );
-            crate::export::obsidian::overwrite_note(std::path::Path::new(sup_path), &new_sup)?;
-            // Same conditional append-side baseline refresh as the source stamp above.
-            refresh_meeting_note_exported_hash(
-                state,
-                &row.superseding_meeting_id,
-                &current_sup,
-                &new_sup,
-            )?;
-        }
-
-        // APPLIED is the LAST write — flipped only after the note(s) are safely stamped.
-        state
-            .db
-            .mark_supersession_applied(id, &chrono::Utc::now().to_rfc3339())?;
-        applied += 1;
-    }
-    tracing::info!(target: "retruth", applied, skipped_sealed, "supersessions applied");
-    Ok(ApplyResult {
-        applied,
-        skipped_sealed,
-    })
-}
-
-/// Resolve the PRISTINE (pre-batch) bytes of a note file from the per-apply-call cache: on the first
-/// touch of a path this batch, read + UTF-8-validate the file and cache it; every later touch reuses
-/// the cached bytes. This is what makes all rows sharing a note carry identical pristine pre-images
-/// (so a multi-row undo restores each file once, byte-identical). Refusing a non-UTF-8 file here —
-/// before any write — keeps the stamp all-or-nothing.
-fn pristine_note_bytes(
-    cache: &mut std::collections::HashMap<String, Vec<u8>>,
-    path: &str,
-) -> Result<Vec<u8>, AppError> {
-    if let Some(b) = cache.get(path) {
-        return Ok(b.clone());
-    }
-    let bytes =
-        std::fs::read(path).map_err(|e| AppError::Export(format!("read note failed: {e}")))?;
-    String::from_utf8(bytes.clone())
-        .map_err(|_| AppError::Export("note is not valid UTF-8".into()))?;
-    cache.insert(path.to_string(), bytes.clone());
-    Ok(bytes)
-}
-
-/// UNDO the given applied supersessions: restore each stamped note's byte-exact pre-image (atomic
-/// overwrite) and clear the row's applied state + pre-images. A row that isn't applied is a no-op. A
-/// note whose folder sealed since apply is SKIPPED (never re-materialize plaintext into a locked
-/// folder) — the sealed content will return WITH the stamp on unlock.
-#[tauri::command]
-pub fn undo_supersessions(state: State<'_, AppState>, ids: Vec<String>) -> Result<(), AppError> {
-    undo_supersessions_inner(state.inner(), &ids)
-}
-
-pub(crate) fn undo_supersessions_inner(state: &AppState, ids: &[String]) -> Result<(), AppError> {
-    // Seal-vs-write TOCTOU (Phase-0 lock-review follow-up): same guard as
-    // `apply_supersessions_inner` — the folder-open checks below and the restore writes must not
-    // interleave with a concurrent lock/relock.
-    let _lifecycle = lifecycle_guard(state);
-    let undo_set: std::collections::HashSet<&str> = ids.iter().map(String::as_str).collect();
-
-    // Collect the DISTINCT affected note files touched by the UNDO SET (`path -> (meeting_id,
-    // pristine pre-image)`). All rows in one heal sharing a note carry IDENTICAL pristine pre-images
-    // (see apply's per-path cache), and `path ↔ meeting` is 1:1. Only files whose folder is still
-    // OPEN are collected/rewritten (never re-materialize plaintext into a sealed folder — a sealed
-    // note's stamp rides inside its sealed content already; and purge-on-seal has already dropped any
-    // supersession referencing a sealed meeting).
-    let mut affected: std::collections::HashMap<String, (String, Vec<u8>)> =
-        std::collections::HashMap::new();
-    let mut to_clear: Vec<&String> = Vec::new();
-    for id in ids {
-        let Some(row) = state.db.get_supersession(id)? else {
-            continue;
-        };
-        if row.applied_at.is_none() {
-            continue; // nothing applied to undo.
-        }
-        if let Some(pre) = &row.source_pre_image {
-            if !folder_locked_on_disk(state, &row.source_meeting_id)? {
-                if let Some((path, _)) = note_file_for(state, &row.source_meeting_id)? {
-                    affected
-                        .entry(path)
-                        .or_insert_with(|| (row.source_meeting_id.clone(), pre.clone()));
-                }
-            }
-        }
-        if let Some(pre) = &row.superseding_pre_image {
-            if !folder_locked_on_disk(state, &row.superseding_meeting_id)? {
-                if let Some((path, _)) = note_file_for(state, &row.superseding_meeting_id)? {
-                    affected
-                        .entry(path)
-                        .or_insert_with(|| (row.superseding_meeting_id.clone(), pre.clone()));
-                }
-            }
-        }
-        to_clear.push(id);
-    }
-
-    // For each affected file: rebuild it as pristine + the stamps of every SURVIVOR — a supersession
-    // that touches THIS note's meeting, is NOT in the undo set, and remains applied. Because the
-    // callout appends are idempotent + order-independent across distinct supersessions, replaying the
-    // survivors reconstructs the exact on-disk state that matches the DB. A FULL undo (no survivors)
-    // collapses to a plain pristine restore. This closes the partial-undo desync where restoring a
-    // shared file to pristine silently stripped a still-applied sibling's on-disk stamp.
-    for (path, (meeting_id, pristine)) in &affected {
-        let mut text = String::from_utf8(pristine.clone())
-            .map_err(|_| AppError::Export("stored pre-image is not valid UTF-8".into()))?;
-        for s in state.db.supersessions_touching_meeting(meeting_id)? {
-            if undo_set.contains(s.id.as_str()) || s.applied_at.is_none() {
-                continue; // being undone, or not currently applied → no stamp to replay.
-            }
-            // Reproduce this survivor's stamp with its ORIGINAL date (the day it was applied) so the
-            // replay byte-matches the original append.
-            let date = s
-                .applied_at
-                .as_deref()
-                .and_then(|a| a.split('T').next())
-                .unwrap_or("")
-                .to_string();
-            if &s.source_meeting_id == meeting_id {
-                // This file is the SURVIVOR's SOURCE note → re-append its `[!superseded]` callout,
-                // reproducing the `· see [[…]]` link exactly as apply did (open superseding only).
-                let sup_stem = superseding_link_stem(state, &s)?;
-                text = crate::export::obsidian::append_supersession_callout(
-                    &text,
-                    &date,
-                    &s.predicate,
-                    &s.old_value,
-                    &s.new_value,
-                    sup_stem.as_deref(),
-                );
-            }
-            if &s.superseding_meeting_id == meeting_id {
-                // This file is the SURVIVOR's SUPERSEDING note → re-append its `[!supersedes]`
-                // backlink referencing the survivor's SOURCE stem.
-                if let Some((_, src_stem)) = note_file_for(state, &s.source_meeting_id)? {
-                    text = crate::export::obsidian::append_supersedes_callout(
-                        &text,
-                        &date,
-                        &s.predicate,
-                        &s.old_value,
-                        &s.new_value,
-                        &src_stem,
-                    );
-                }
-            }
-        }
-        // Export-collision guard: the undo rebuild is a FULL overwrite from the stored pre-image
-        // (+ survivor replays), so an external edit made since apply is preserved as a sibling
-        // first, and the baseline re-stamped from the rebuilt content. The no-note-row branch is
-        // dead-defensive (`affected` is only ever keyed via `note_file_for`, which requires a
-        // note row) — it still routes through the ONE guarded overwrite (Phase-0 follow-up): a
-        // missing row reads a NULL baseline (grandfathered, no sibling) and its hash re-stamp
-        // updates zero rows, so behavior is identical while the invariant "every full overwrite
-        // of an exported note goes through the guard" holds structurally.
-        let provider_id = state
-            .db
-            .get_latest_note_for_meeting(meeting_id)?
-            .map(|n| n.provider_id)
-            .unwrap_or_default();
-        overwrite_exported_note_guarded(state, meeting_id, &provider_id, path, &text)?;
-    }
-
-    // Clear the applied state on the undone rows ONLY (pre-images dropped); survivors stay applied.
-    for id in to_clear {
-        state.db.clear_supersession_applied(id)?;
-    }
-    Ok(())
-}
-
-/// The `· see [[stem]]` link stem for a supersession's SOURCE-side callout: the superseding note's
-/// file-stem when that note is open-on-disk + unlocked (exactly the apply-time condition), else
-/// `None` (never leak a sealed meeting's title). Mirrors the apply path so a survivor replay
-/// reproduces the original callout.
-fn superseding_link_stem(
-    state: &AppState,
-    s: &crate::storage::models::SupersessionRow,
-) -> Result<Option<String>, AppError> {
-    let open = meeting_is_unlocked(state, &s.superseding_meeting_id)?
-        && !folder_locked_on_disk(state, &s.superseding_meeting_id)?;
-    if !open {
-        return Ok(None);
-    }
-    Ok(note_file_for(state, &s.superseding_meeting_id)?.map(|(_, stem)| stem))
 }
 
 // ── Vault Audit v1 — deterministic vault-health inbox (see `crate::audit`) ──────────────────────
@@ -5225,236 +3589,6 @@ pub async fn entity_dossier(
 #[path = "tests/entity_dossier_tests.rs"]
 mod entity_dossier_tests;
 
-/// Generate a Weekly Vault Digest synthesizing meetings from the last `days` days; writes it
-/// into the vault's Digests/ folder and returns the markdown + path.
-#[tauri::command]
-pub async fn generate_digest(
-    state: State<'_, AppState>,
-    days: i64,
-) -> Result<DigestResult, AppError> {
-    let days = days.clamp(1, 90);
-    let config = {
-        state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?
-            .clone()
-    };
-    let cutoff = (chrono::Utc::now() - chrono::Duration::days(days)).to_rfc3339();
-    // Budget on the NOTES-role provider's RESOLVED connection — the corpus egresses to it
-    // (identical to `provider_id` while role keys are absent).
-    let notes_conn =
-        crate::summarize::roles::provider_target(crate::summarize::roles::Role::Notes, &config)
-            .connection;
-    let budget = if notes_conn == "ollama" {
-        4_000
-    } else {
-        80_000
-    };
-    // Finding 2 + BLK-2b: build the cloud corpus from VISIBLE meetings + VISIBLE notes only, so a
-    // sealed-and-not-unlocked meeting's TITLE (the `### [[title]]` header) AND markdown never leave
-    // the device. `list_meetings_visible` + `get_note_if_visible` push the session unlock set
-    // through the same predicate as MCP — correctness no longer depends on at-rest blanking.
-    let unlocked = unlocked_snapshot(state.inner())?;
-    let mut corpus = String::new();
-    let mut count = 0usize;
-    for m in state.db.list_meetings_visible(300, &unlocked)? {
-        if m.started_at.as_str() < cutoff.as_str() {
-            continue;
-        }
-        if corpus.len() >= budget {
-            break;
-        }
-        let Some(note) = state.db.get_note_if_visible(&m.id, &unlocked)? else {
-            continue;
-        };
-        let title = m.title.clone().unwrap_or_else(|| "(untitled)".to_string());
-        let date = m
-            .started_at
-            .split(['T', ' '])
-            .next()
-            .unwrap_or("")
-            .to_string();
-        let header = format!("\n\n### [[{title}]] · {date}\n");
-        let remaining = budget.saturating_sub(corpus.len() + header.len());
-        if remaining < 200 {
-            break;
-        }
-        corpus.push_str(&header);
-        corpus.push_str(&note.markdown.chars().take(remaining).collect::<String>());
-        count += 1;
-    }
-    if count == 0 {
-        return Err(AppError::InvalidArg(format!(
-            "no summarized meetings in the last {days} days"
-        )));
-    }
-    let range_label = format!("the last {days} days");
-    let provider = crate::summarize::provider_for(crate::summarize::roles::Role::Notes, &config, &state.heavy_inference)?;
-    let (system, user) =
-        crate::summarize::digest::build_digest_prompt(&corpus, &range_label, &config.note_language);
-    let markdown = provider.complete(&system, &user).await?;
-
-    let exported_path = match config.vault_path.as_deref().filter(|p| !p.is_empty()) {
-        Some(vault) => {
-            let now = chrono::Utc::now().to_rfc3339();
-            crate::export::write_note(
-                std::path::Path::new(vault),
-                Some("Digests"),
-                "Weekly Digest",
-                &now,
-                &markdown,
-            )
-            .ok()
-            .map(|p| p.to_string_lossy().to_string())
-        }
-        None => None,
-    };
-    Ok(DigestResult {
-        markdown,
-        exported_path,
-    })
-}
-
-// ── Brain v2 L5 — scheduled briefs (schedule CRUD + propose-accept runs) ─────────────────────────
-
-/// All brief schedules (config rows — labels, timing, hints; no meeting content).
-#[tauri::command]
-pub fn list_brief_schedules(
-    state: State<'_, AppState>,
-) -> Result<Vec<crate::storage::models::BriefSchedule>, AppError> {
-    state.db.list_brief_schedules()
-}
-
-/// Validate a brief schedule's user-editable fields (shared by create + update).
-fn validate_brief_schedule(s: &crate::storage::models::BriefSchedule) -> Result<(), AppError> {
-    if s.label.trim().is_empty() {
-        return Err(AppError::InvalidArg("brief label is empty".into()));
-    }
-    if let Some(d) = s.day_of_week {
-        if !(0..=6).contains(&d) {
-            return Err(AppError::InvalidArg(
-                "day_of_week must be 0 (Monday) … 6 (Sunday)".into(),
-            ));
-        }
-    }
-    if !(0..=23).contains(&s.hour_local) {
-        return Err(AppError::InvalidArg("hour must be 0…23".into()));
-    }
-    if !(0..=59).contains(&s.minute_local) {
-        return Err(AppError::InvalidArg("minute must be 0…59".into()));
-    }
-    if !(1..=90).contains(&s.scope_days) {
-        return Err(AppError::InvalidArg("scope_days must be 1…90".into()));
-    }
-    Ok(())
-}
-
-/// Create one brief schedule. `day_of_week`: 0 = Monday … 6 = Sunday, `None` = daily. The runner
-/// (`crate::brief_runner`) fires it at most once per local day; the first fire is the first 60s
-/// tick at/after `hour:minute` local.
-#[tauri::command]
-pub fn create_brief_schedule(
-    state: State<'_, AppState>,
-    label: String,
-    day_of_week: Option<i64>,
-    hour_local: i64,
-    minute_local: i64,
-    scope_days: Option<i64>,
-    prompt_hint: Option<String>,
-) -> Result<crate::storage::models::BriefSchedule, AppError> {
-    let schedule = crate::storage::models::BriefSchedule {
-        id: uuid::Uuid::new_v4().simple().to_string(),
-        label: label.trim().to_string(),
-        day_of_week,
-        hour_local,
-        minute_local,
-        scope_days: scope_days.unwrap_or(7),
-        prompt_hint: prompt_hint.map(|h| h.trim().to_string()).filter(|h| !h.is_empty()),
-        enabled: true,
-        last_run_at: None,
-        created_at: chrono::Utc::now().to_rfc3339(),
-    };
-    validate_brief_schedule(&schedule)?;
-    state.db.insert_brief_schedule(&schedule)?;
-    Ok(schedule)
-}
-
-/// Update one brief schedule's editable fields (label / timing / window / hint / enabled).
-#[tauri::command]
-pub fn update_brief_schedule(
-    state: State<'_, AppState>,
-    schedule: crate::storage::models::BriefSchedule,
-) -> Result<(), AppError> {
-    validate_brief_schedule(&schedule)?;
-    state.db.update_brief_schedule(&schedule)
-}
-
-/// Delete one brief schedule AND its staged runs.
-#[tauri::command]
-pub fn delete_brief_schedule(
-    state: State<'_, AppState>,
-    schedule_id: String,
-) -> Result<(), AppError> {
-    state.db.delete_brief_schedule(&schedule_id)
-}
-
-/// The PENDING (proposed, not yet accepted/dismissed) brief runs — the FE's proposal cards.
-/// `note_md` was synthesized by the runner from VISIBLE-ONLY content (empty unlock set — the
-/// consolidation-job discipline), so it cannot contain sealed content AT synthesis time; and a
-/// meeting sealed AFTER the proposal purges its pending runs inside the seal tx
-/// (`Db::purge_pending_brief_runs_tx` — the lock-security LEAK fix, 2026-07-10), so a row this
-/// returns never paraphrases a currently-sealed meeting. That pair is what makes this read safe
-/// without a per-meeting gate (documented posture, see `crate::brief_runner` + `migrate_briefs`).
-#[tauri::command]
-pub fn list_brief_runs(
-    state: State<'_, AppState>,
-) -> Result<Vec<crate::storage::models::BriefRun>, AppError> {
-    state.db.list_pending_brief_runs()
-}
-
-/// ACCEPT a proposed brief: export its markdown to `<vault>/Briefs/` (atomic write) and CONSUME
-/// the staged `note_md` (the vault `.md` becomes the only copy). Returns the exported path.
-#[tauri::command]
-pub fn accept_brief(state: State<'_, AppState>, run_id: String) -> Result<String, AppError> {
-    let run = state
-        .db
-        .get_brief_run(&run_id)?
-        .ok_or_else(|| AppError::InvalidArg(format!("no brief run {run_id}")))?;
-    if run.status != "pending" {
-        return Err(AppError::InvalidArg("this brief was already handled".into()));
-    }
-    if run.note_md.trim().is_empty() {
-        return Err(AppError::InvalidArg("this brief has no content".into()));
-    }
-    let vault = vault_path(state.inner())
-        .ok_or_else(|| AppError::InvalidArg("set an Obsidian vault first (Settings)".into()))?;
-    let label = state
-        .db
-        .list_brief_schedules()?
-        .into_iter()
-        .find(|s| s.id == run.schedule_id)
-        .map(|s| s.label)
-        .unwrap_or_else(|| "Brief".to_string());
-    let path = crate::export::write_note(
-        std::path::Path::new(&vault),
-        Some("Briefs"),
-        &label,
-        &run.proposed_at,
-        &run.note_md,
-    )?;
-    state
-        .db
-        .accept_brief_run(&run_id, &chrono::Utc::now().to_rfc3339())?;
-    Ok(path.to_string_lossy().to_string())
-}
-
-/// DISMISS a proposed brief: the staged row (markdown included) is deleted outright.
-#[tauri::command]
-pub fn dismiss_brief(state: State<'_, AppState>, run_id: String) -> Result<(), AppError> {
-    state.db.delete_brief_run(&run_id)
-}
-
 // ── Brain v2 L5 — MCP server config (list/add/remove + per-server consent + test) ───────────────
 //
 // The list/add/remove + per-server consent commands were extracted verbatim to `commands/mcp.rs`
@@ -5525,76 +3659,6 @@ pub fn topic_threads(state: State<'_, AppState>) -> Result<Vec<TopicThread>, App
         });
     }
     Ok(crate::summarize::threads::build_threads(&input))
-}
-
-/// Best-effort: the soonest macOS Calendar event in the next 60 minutes (title only). Returns
-/// None if Calendar access is denied or there's nothing upcoming — never errors the UI.
-#[tauri::command]
-pub async fn next_calendar_event() -> Result<Option<CalendarEvent>, AppError> {
-    let script = r#"set now to (current date)
-set laterT to now + (60 * minutes)
-set out to ""
-try
-  tell application "Calendar"
-    repeat with c in calendars
-      repeat with e in (every event of c whose start date is greater than or equal to now and start date is less than or equal to laterT)
-        set out to out & (summary of e) & linefeed
-      end repeat
-    end repeat
-  end tell
-end try
-return out"#;
-    let res = tokio::task::spawn_blocking(move || {
-        std::process::Command::new("osascript")
-            .arg("-e")
-            .arg(script)
-            .output()
-    })
-    .await;
-    let stdout = match res {
-        Ok(Ok(o)) if o.status.success() => String::from_utf8_lossy(&o.stdout).into_owned(),
-        _ => return Ok(None),
-    };
-    let title = stdout
-        .lines()
-        .map(str::trim)
-        .find(|l| !l.is_empty())
-        .map(str::to_string);
-    Ok(title.map(|title| CalendarEvent { title, start: None }))
-}
-
-/// CALENDAR source (local, zero-OAuth, on-device): list the user's events in a window around now
-/// via the bundled `meetnotes-calendar` EventKit sidecar — title, attendees, agenda. GRACEFUL on
-/// every failure: sidecar missing / Calendar permission denied / timeout / malformed output →
-/// an empty list, never an error, never a block. No network egress: reading the local calendar
-/// stays on device.
-#[tauri::command]
-pub async fn list_calendar_events(app: AppHandle) -> Result<Vec<CalendarEventFull>, AppError> {
-    // Default window: now-1h .. now+12h (60 back, 720 forward minutes).
-    Ok(crate::calendar::fetch_events(&app, 60, 720).await)
-}
-
-/// Build a compact [`CalendarContext`] (title + attendees + agenda) for one event so the existing
-/// pre-meeting brief / note pre-analysis can consume it (the brain already takes context). Looks
-/// the event up by id in the same window the sidecar surfaces. Returns `None` if the event isn't
-/// found (expired from the window, or Calendar access denied) — never an error.
-///
-/// IMPORTANT: the returned text is on-device context. If it is later fed to a CLOUD provider it
-/// MUST ride the existing `make_provider` redaction firewall + consent (the same path the
-/// transcript takes) — this command opens NO new egress path.
-#[tauri::command]
-pub async fn calendar_context_for(
-    app: AppHandle,
-    event_id: String,
-) -> Result<Option<CalendarContext>, AppError> {
-    if event_id.trim().is_empty() {
-        return Err(AppError::InvalidArg("event_id is empty".into()));
-    }
-    let events = crate::calendar::fetch_events(&app, 60, 720).await;
-    Ok(events
-        .iter()
-        .find(|e| e.id == event_id)
-        .map(CalendarContext::from_event))
 }
 
 /// Read current config (settings table), without secrets.
@@ -6068,222 +4132,6 @@ fn dto_to_config(d: AppConfigDto, current: &AppConfig) -> AppConfig {
 
 
 
-/// DTO for a single model returned by `list_gateway_models`.
-///
-/// Shape: `{ "id": "gpt-4o" }` (camelCase). The FE populates the model picker from this.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GatewayModelDto {
-    pub id: String,
-}
-
-/// Fetch the model catalog from the configured AI Gateway (`GET {base}/v1/models`) and return
-/// the raw list of model ids. Shared core of `list_gateway_models` and `list_models("gateway")`.
-///
-/// Inbound-only: this call sends NO meeting content — only an optional `Authorization: Bearer`
-/// header carrying the stored gateway key. It therefore does NOT need the redaction firewall or
-/// the cloud-egress consent gate.
-///
-/// Returns `AppError::InvalidArg` when no gateway base URL is configured.
-async fn gateway_model_ids(state: &AppState) -> Result<Vec<String>, AppError> {
-    let (base_url, model, api_key) = {
-        let config = state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?;
-        let base_url = config.gateway_base_url.clone();
-        let model = config.gateway_model.clone();
-        // R3: resolve the gateway key; never falls back to the Anthropic key.
-        let api_key = secrets::get_secret(GATEWAY_KEY_ACCOUNT).ok().flatten();
-        (base_url, model, api_key)
-    };
-
-    if base_url.trim().is_empty() {
-        return Err(AppError::InvalidArg(
-            "no gateway base URL configured — set it in Settings before fetching the model catalog"
-                .into(),
-        ));
-    }
-
-    let provider = crate::summarize::gateway::OpenAiCompatProvider::new(base_url, model, api_key)?;
-    provider.list_models().await
-}
-
-/// Fetch the model catalog from the configured AI Gateway (`GET {base}/v1/models`) and return
-/// the list of model ids the gateway exposes. Thin DTO wrapper over [`gateway_model_ids`] —
-/// kept until the FE swaps to the unified `list_models("gateway")`.
-#[tauri::command]
-pub async fn list_gateway_models(
-    state: State<'_, AppState>,
-) -> Result<Vec<GatewayModelDto>, AppError> {
-    let ids = gateway_model_ids(&state).await?;
-    Ok(ids.into_iter().map(|id| GatewayModelDto { id }).collect())
-}
-
-/// Model ids for the connections whose catalog is static, compile-time data — pure so it is
-/// unit-testable without state or network. `None`-network connections only:
-///   - `"claude_code"` / `"anthropic"` → the curated [`crate::summarize::provider::CLAUDE_MODELS`],
-///   - `"local"` → the on-device `reason::BRAIN_MODELS` registry ids,
-///   - `"off"` → empty (a valid connection that runs no models),
-///   - anything else → `AppError::InvalidArg`.
-fn static_connection_models(connection: &str) -> Result<Vec<String>, AppError> {
-    match connection {
-        "claude_code" | "anthropic" => Ok(crate::summarize::provider::CLAUDE_MODELS
-            .iter()
-            .map(|id| id.to_string())
-            .collect()),
-        "local" => Ok(crate::reason::BRAIN_MODELS
-            .iter()
-            .map(|m| m.id.to_string())
-            .collect()),
-        "off" => Ok(vec![]),
-        other => Err(AppError::InvalidArg(format!(
-            "unknown connection '{other}' — expected claude_code, anthropic, ollama, gateway, local, or off"
-        ))),
-    }
-}
-
-/// Unified model catalog for a connection — the ONE source of truth behind the FE per-role
-/// model dropdowns. Dispatch by `connection`:
-///   - `"gateway"` → `GET {gateway_base_url}/v1/models` (shares [`gateway_model_ids`] with
-///     `list_gateway_models`),
-///   - `"ollama"` → `GET {ollama_base_url}/api/tags`, model names only,
-///   - `"claude_code"` / `"anthropic"` / `"local"` / `"off"` → static lists
-///     (see [`static_connection_models`]),
-///   - anything else → `AppError::InvalidArg`.
-///
-/// Inbound-only on every arm: NO meeting content is sent — at most an `Authorization: Bearer`
-/// header to the configured gateway — so no redaction firewall / consent gate is needed (the
-/// `list_gateway_models` precedent). No content read → no lock-model surface.
-#[tauri::command]
-pub async fn list_models(
-    state: State<'_, AppState>,
-    connection: String,
-) -> Result<Vec<String>, AppError> {
-    match connection.as_str() {
-        "gateway" => gateway_model_ids(&state).await,
-        "ollama" => {
-            let base_url = {
-                let config = state
-                    .config
-                    .lock()
-                    .map_err(|_| AppError::Config("config mutex poisoned".into()))?;
-                config.ollama_base_url.clone()
-            };
-            // OllamaProvider::new normalizes an empty base URL to the localhost default and
-            // trims trailing slashes; the model arg is unused for a catalog fetch.
-            crate::summarize::ollama::OllamaProvider::new(base_url, String::new())
-                .list_models()
-                .await
-        }
-        other => static_connection_models(other),
-    }
-}
-
-/// DTO returned by `gateway_health`.
-///
-/// Shape: `{ "reachable": true, "modelCount": 6 }` (camelCase, matches the FE `GatewayHealth`
-/// type). `reachable: false` with `model_count: 0` means the gateway is unreachable or not
-/// configured — the FE renders a red dot.
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GatewayHealthDto {
-    pub reachable: bool,
-    pub model_count: u32,
-}
-
-/// Probe the configured AI Gateway for reachability and (optionally) catalog size.
-///
-/// Uses `OpenAiCompatProvider::probe()` which sends a GET to the models endpoint when one
-/// exists, or to the chat endpoint for custom routes (a GET to a POST-only route returns
-/// 4xx/405 but proves the server is reachable — no LLM call, no cost). Any HTTP response
-/// (any status) → `reachable: true`; a transport failure → `reachable: false`.
-///
-/// This command NEVER returns an `Err` variant so the FE health dot always gets a clean value.
-/// Inbound-only: sends NO meeting content, only an optional `Authorization: Bearer` header. Does
-/// NOT need the redaction firewall or the consent gate (same rationale as `list_gateway_models`).
-#[tauri::command]
-pub async fn gateway_health(state: State<'_, AppState>) -> Result<GatewayHealthDto, AppError> {
-    let (base_url, model, api_key) = {
-        let config = match state.config.lock() {
-            Ok(c) => c,
-            Err(_) => {
-                return Ok(GatewayHealthDto {
-                    reachable: false,
-                    model_count: 0,
-                })
-            }
-        };
-        let base_url = config.gateway_base_url.clone();
-        let model = config.gateway_model.clone();
-        // R3: resolve the gateway key; never falls back to the Anthropic key.
-        let api_key = secrets::get_secret(GATEWAY_KEY_ACCOUNT).ok().flatten();
-        (base_url, model, api_key)
-    };
-
-    if base_url.trim().is_empty() {
-        // Not configured → degrade silently.
-        return Ok(GatewayHealthDto {
-            reachable: false,
-            model_count: 0,
-        });
-    }
-
-    let provider =
-        match crate::summarize::gateway::OpenAiCompatProvider::new(base_url, model, api_key) {
-            Ok(p) => p,
-            Err(_) => {
-                return Ok(GatewayHealthDto {
-                    reachable: false,
-                    model_count: 0,
-                })
-            }
-        };
-
-    // probe() never returns Err — degrades to (false, 0) on transport failure.
-    let (reachable, model_count) = provider.probe().await;
-    Ok(GatewayHealthDto {
-        reachable,
-        model_count,
-    })
-}
-
-// ── Egress ledger DTOs (Phase 6) ────────────────────────────────────────────────────────────────
-//
-// All structs are `camelCase` on the wire (matches the FE `EgressLedger` / `EgressRow` types in
-// `core/models.ts`). Carries ONLY counts, ids, labels, and token/byte numbers — no content (§8).
-
-/// availability() fan-out across all three providers for the Settings UI.
-#[tauri::command]
-pub async fn provider_statuses(
-    state: State<'_, AppState>,
-) -> Result<Vec<ProviderStatus>, AppError> {
-    use crate::summarize::provider::Availability;
-
-    let config: AppConfig = {
-        let guard = state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?;
-        guard.clone()
-    };
-
-    let providers = all_providers(&config);
-    let mut out = Vec::with_capacity(providers.len());
-    for p in providers {
-        let (available, reason) = match p.availability().await {
-            Availability::Available => (true, None),
-            Availability::Unavailable { reason } => (false, Some(reason)),
-        };
-        out.push(ProviderStatus {
-            id: p.id().to_string(),
-            available,
-            reason,
-        });
-    }
-    Ok(out)
-}
-
 /// Re-run summarize+export for an existing meeting with the configured provider, reusing
 /// the meeting's stored transcript segments (Detail "re-summarize"/"re-export" seed —
 /// wired in P0, UI optional).
@@ -6594,184 +4442,6 @@ pub(crate) fn masked_detail(meeting: Meeting) -> MeetingDetailDto {
     }
 }
 
-/// Download the Whisper model matching the chosen size + language (multilingual unless
-/// English is selected) from the whisper.cpp HuggingFace mirror into the app models dir if
-/// missing; returns its path. No-op (returns the existing path) when already present. Emits
-/// [`crate::events::EVENT_MODEL_DOWNLOAD`] progress (throttled) so the FE can show a progress bar.
-#[tauri::command]
-pub async fn download_model(
-    app: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<String, AppError> {
-    let (configured, size, language) = {
-        let c = state
-            .config
-            .lock()
-            .map_err(|_| AppError::Config("config mutex poisoned".into()))?;
-        (
-            c.whisper_model_path.clone(),
-            c.model_size.clone(),
-            c.language.clone().unwrap_or_default(),
-        )
-    };
-    let p = configured.as_deref().map(std::path::Path::new);
-
-    // Throttle progress events to roughly every 8 MB so a multi-GB download doesn't flood the FE.
-    const EMIT_EVERY: u64 = 8 * 1024 * 1024;
-    let mut last_emit: u64 = 0;
-    let path = crate::transcribe::ensure_model(p, &size, &language, |downloaded, total| {
-        if downloaded - last_emit >= EMIT_EVERY {
-            last_emit = downloaded;
-            let _ = app.emit(
-                crate::events::EVENT_MODEL_DOWNLOAD,
-                crate::events::ModelDownloadPayload {
-                    downloaded,
-                    total,
-                    done: false,
-                },
-            );
-        }
-    })
-    .await?;
-
-    let _ = app.emit(
-        crate::events::EVENT_MODEL_DOWNLOAD,
-        crate::events::ModelDownloadPayload {
-            downloaded: 0,
-            total: None,
-            done: true,
-        },
-    );
-    Ok(path.to_string_lossy().to_string())
-}
-
-
-/// Download the OPTIONAL parakeet live-ASR engine's four int8 models (~600 MB) from the csukuangfj
-/// sherpa-onnx HF mirror into `<models_dir>/parakeet-tdt-0.6b-v3-int8` if missing (atomic per-file
-/// `.part` → rename; a file already present is skipped). Emits the SAME
-/// [`crate::events::EVENT_MODEL_DOWNLOAD`] progress (throttled) as the whisper download so the FE
-/// reuses one progress bar. No-op when all four are already present.
-#[tauri::command]
-pub async fn download_parakeet_models(app: AppHandle) -> Result<(), AppError> {
-    // Throttle progress events to roughly every 8 MB (parity with `download_model`).
-    const EMIT_EVERY: u64 = 8 * 1024 * 1024;
-    let mut last_emit: u64 = 0;
-    crate::transcribe::model::ensure_parakeet_models(|downloaded, total| {
-        if downloaded - last_emit >= EMIT_EVERY {
-            last_emit = downloaded;
-            let _ = app.emit(
-                crate::events::EVENT_MODEL_DOWNLOAD,
-                crate::events::ModelDownloadPayload {
-                    downloaded,
-                    total,
-                    done: false,
-                },
-            );
-        }
-    })
-    .await?;
-
-    let _ = app.emit(
-        crate::events::EVENT_MODEL_DOWNLOAD,
-        crate::events::ModelDownloadPayload {
-            downloaded: 0,
-            total: None,
-            done: true,
-        },
-    );
-    Ok(())
-}
-
-/// Download the on-device brain model identified by `model_id` (from the curated registry) into the
-/// shared models dir if missing; returns its path. No-op (returns the existing path) when already
-/// present. Unknown id ⇒ `AppError::InvalidArg`. INBOUND ONLY — fetches a model file and sends NO
-/// meeting content (no egress). Emits [`crate::events::EVENT_BRAIN_DOWNLOAD`] progress events
-/// (throttled). The downloaded file is NOT loaded here — wiring the brain is a later step.
-/// (The registry-lookup core `brain_download_target` lives in `commands/model_perf.rs`.)
-#[tauri::command]
-pub async fn download_brain_model(app: AppHandle, model_id: String) -> Result<String, AppError> {
-    let (url, dest) = brain_download_target(&model_id)?;
-
-    if dest.is_file() {
-        return Ok(dest.to_string_lossy().to_string());
-    }
-
-    // The pinned integrity hash for this model (None until the registry entry is pinned — the
-    // downloader then verifies before promoting the file, or warns + skips when None).
-    let expected_sha = crate::reason::brain_model_by_id(&model_id).and_then(|m| m.sha256);
-
-    // Throttle progress events to roughly every 8 MB so a multi-GB download doesn't flood the FE.
-    const EMIT_EVERY: u64 = 8 * 1024 * 1024;
-    let mut last_emit: u64 = 0;
-    crate::reason::download_brain_model(url, &dest, expected_sha, |downloaded, total| {
-        if downloaded - last_emit >= EMIT_EVERY {
-            last_emit = downloaded;
-            let _ = app.emit(
-                crate::events::EVENT_BRAIN_DOWNLOAD,
-                crate::events::BrainDownloadPayload {
-                    downloaded,
-                    total,
-                    done: false,
-                },
-            );
-        }
-    })
-    .await?;
-
-    let _ = app.emit(
-        crate::events::EVENT_BRAIN_DOWNLOAD,
-        crate::events::BrainDownloadPayload {
-            downloaded: 0,
-            total: None,
-            done: true,
-        },
-    );
-    Ok(dest.to_string_lossy().to_string())
-}
-
-/// Download the multilingual-e5-small model (3 HF files) into the shared models dir, INBOUND-ONLY,
-/// emitting [`crate::events::EVENT_EMBED_DOWNLOAD`] progress (throttled per file). Sends NO meeting
-/// content (no egress). The downloaded model is NOT loaded here — it is picked up lazily by
-/// `embed::active_embedder` on the next embed (selected on model presence). Returns the model dir.
-#[tauri::command]
-pub async fn download_embed_model(app: AppHandle) -> Result<String, AppError> {
-    let file_count = crate::embed::EMBED_MODEL_FILES.len();
-    // Throttle progress to roughly every 2 MB so the (small) model download doesn't flood the FE.
-    const EMIT_EVERY: u64 = 2 * 1024 * 1024;
-    let mut last_emit: u64 = 0;
-    let mut last_index: usize = usize::MAX;
-    let dir = crate::embed::download_embed_model(|file_index, downloaded, total| {
-        // Always emit on a file boundary; otherwise throttle by bytes.
-        if file_index != last_index || downloaded - last_emit >= EMIT_EVERY {
-            last_index = file_index;
-            last_emit = downloaded;
-            let _ = app.emit(
-                crate::events::EVENT_EMBED_DOWNLOAD,
-                crate::events::EmbedDownloadPayload {
-                    file_index,
-                    file_count,
-                    downloaded,
-                    total,
-                    done: false,
-                },
-            );
-        }
-    })
-    .await?;
-
-    let _ = app.emit(
-        crate::events::EVENT_EMBED_DOWNLOAD,
-        crate::events::EmbedDownloadPayload {
-            file_index: file_count,
-            file_count,
-            downloaded: 0,
-            total: None,
-            done: true,
-        },
-    );
-    Ok(dir.to_string_lossy().to_string())
-}
-
 /// Result of [`reindex_embeddings`]. `status` is `"model_missing"` when the real e5 model is absent
 /// (no indexing was attempted — re-indexing with the deterministic STUB embedder would poison the
 /// index with garbage vectors, worse than nothing), else `"indexed"`. On `"indexed"`, `indexed` is
@@ -6782,71 +4452,6 @@ pub struct ReindexResult {
     pub status: String,
     pub indexed: usize,
     pub total: usize,
-}
-
-/// brain2 RAG — BACKFILL the semantic vector index for ALL VISIBLE meetings (the one-shot the user
-/// runs after turning `semantic_search_enabled` on, or after installing the e5 model so the old
-/// STUB-embedded chunks get replaced by real e5 vectors).
-///
-/// GATING (lock-model): the corpus is exactly `list_meetings_visible(unlocked)` — a sealed-and-not-
-/// session-unlocked meeting is NEVER returned, so its plaintext is never chunked/embedded, and its
-/// chunks STAY purged (the seal already purged them; we don't touch them). For each visible meeting
-/// we re-fetch its note through `get_note_if_visible(unlocked)` (defense-in-depth: skip if the note
-/// is not visible) and call `index_meeting_chunks`, which PURGES-then-reinserts — so any stale stub
-/// vectors are replaced with e5 ones. No new read path: every read routes through `visibility_clause`.
-///
-/// MODEL GUARD: if the real e5 model is absent (`!embed_model_present()` ⇒ `active_embedder` is the
-/// stub), MEETING indexing does nothing and the result is `{ status: "model_missing" }` — re-indexing
-/// with garbage stub vectors is strictly worse than leaving the (old, possibly-stub) chunks alone;
-/// the FE prompts the user to download e5 first via `download_embed_model`. DOCUMENT chunk/FTS
-/// backfill still runs (chunk-only, zero vectors) so keyword retrieval over documents works
-/// regardless of the model.
-///
-/// Emits [`crate::events::EVENT_REINDEX`] `{ done, total }` progress (counts only, NO PII).
-/// EMBED_DIM stays 384 (e5 == stub width) ⇒ NO `vec_chunks` schema migration.
-#[tauri::command]
-pub async fn reindex_embeddings(
-    app: AppHandle,
-    state: State<'_, AppState>,
-) -> Result<ReindexResult, AppError> {
-    // Snapshot the LIVE session unlock set — visibility is evaluated against exactly this set, so a
-    // sealed-not-unlocked folder's meetings are invisible and never indexed.
-    let unlocked = state
-        .unlocked_folders
-        .lock()
-        .map(|g| g.clone())
-        .unwrap_or_default();
-
-    // 2026-07-13 perf audit (MODERATE): this ran `reindex_embeddings_inner` — a vault-wide loop
-    // (up to 100_000 meetings + every visible document) doing real Candle/Metal embed calls —
-    // INLINE on the async command's Tokio worker, the same shape as the original startup-freeze
-    // bug this whole investigation started from (now fixed there via a RAM floor + per-run cap;
-    // the per-meeting embed BATCH SIZE is already safe here too, since index_meeting_chunks/
-    // index_meeting_topic_chunks were sub-batched earlier today). This one is user-TRIGGERED (the
-    // Settings "Reindex" button), not automatic, so the risk is lower, but the fix is the same
-    // shape: move off the async runtime + gate on a RAM floor before starting.
-    if !crate::transcribe::model::topic_backfill_ram_permits_now() {
-        return Err(AppError::Unavailable(
-            "not enough free memory to reindex right now — close some apps and try again".into(),
-        ));
-    }
-    let db = state.db.clone();
-    let heavy_inference = state.heavy_inference.clone();
-    crate::perf::run_heavy(&heavy_inference, move || {
-        reindex_embeddings_inner(
-            &db,
-            &unlocked,
-            crate::embed::embed_model_present(),
-            crate::embed::active_embedder().as_ref(),
-            |done, total| {
-                let _ = app.emit(
-                    crate::events::EVENT_REINDEX,
-                    crate::events::ReindexPayload { done, total },
-                );
-            },
-        )
-    })
-    .await
 }
 
 /// "Powiązane wg znaczenia": the up-to-5 meetings most semantically similar to `meeting_id`.
